% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbForests_CsvToDf.R
\name{AbForests_CsvToDf}
\alias{AbForests_CsvToDf}
\title{Convert list of csvs, to nested list of data.frames}
\usage{
AbForests_CsvToDf(files)
}
\arguments{
\item{files}{a list of csv files. Each csv file may contain information concerning full length heavy and light chain sequences, CDRH3 and CDRL3 sequences, the type of isotype and the transcriptional cluster that corresponds to each of these sequences.}
}
\value{
graphs a list of data.frames. Each data.frame contains 2 columns, one that describes the sequences and the other the type of information (isotype or cluster) is considered in the analysis. All these cases are determined by the user.
}
\description{
CsvToDf converts a list of csv files, which are clone lineages to a list of data.frames.
}
\examples{
\dontrun{
CsvToDf(files)
}
}
\seealso{
AntibodyForest
}
