% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntibodyForest_network_to_pnp_DL.R
\name{AntibodyForest_isotype_filter}
\alias{AntibodyForest_isotype_filter}
\title{Filter function for the AntibodyForest_network_to_pnp function
Input: Dataframe with node features for 1 specific clonotype of 1 specific sample and a vector of character strings, which represent the $VDJ_cgene features one wants to filter}
\usage{
AntibodyForest_isotype_filter(
  node_dfs_single,
  i,
  filter.isotype,
  established_isotypes
)
}
\arguments{
\item{filter.isotype}{Vector of character strings. Should be values found in the $VDJ_cgene column}

\item{node_df_single}{Dataframe of node_features extracted from the igraph objects from the AntibodyForest object - already eventually processed by custom.node.filter}

\item{established.isotypes}{Character string of established VDJ_cgene designations. If the strings in filter.isotype deviate from the established_isotypes, it will give a warning.}
}
\value{
A boolean vector equaling the number of nodes - which nodes pass the filter
}
\description{
Filter function for the AntibodyForest_network_to_pnp function
Input: Dataframe with node features for 1 specific clonotype of 1 specific sample and a vector of character strings, which represent the $VDJ_cgene features one wants to filter
}
