% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntibodyForest_network_to_pnp_DL.R
\name{AntibodyForest_network_filter}
\alias{AntibodyForest_network_filter}
\title{Filter function for the AntibodyForest_network_to_pnp function
Input: AntibodyForest for 1 sample and a (list of) expression(s), which are conditions one wants to filter the clonotypes/networks which are in the ABForest.}
\usage{
AntibodyForest_network_filter(
  AntibodyForest.networks.sample,
  custom.network.filter
)
}
\arguments{
\item{AntibodyForest.networks.sample}{One sample (sample_ids need to be unique) from AntibodyForest.networks, which is a list of clonotypes/igraph objects (AbForest output containing mixcr alignemnt results in network/igraph objects)}

\item{custom.network.filter}{List of expressions, which are to evaluated using eval(). Multiple expressions are "OR'd". If one wishes to "AND" multiple expressions, one has to write that explicitely in one condition}

\item{network.selection}{Boolean vector. Is iterated upon over the conditions and is returned at the end.}

\item{platypus.version}{This function works with "v3" only, there is no need to set this parameter.}
}
\value{
A boolean vector with TRUE at the indexes of the clonotypes that passed all the conditions set in custom.network.filter and FALSE on all the other ones
}
\description{
Filter function for the AntibodyForest_network_to_pnp function
Input: AntibodyForest for 1 sample and a (list of) expression(s), which are conditions one wants to filter the clonotypes/networks which are in the ABForest.
}
