% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntibodyForest_network_to_pnp_MV.R
\name{AntibodyForest_network_to_pnp_MV}
\alias{AntibodyForest_network_to_pnp_MV}
\title{From Networks to PnP Sequences
Extracts from Abforest network function. Initial Version; extracts n most expanded Antibody Sequences per Network from Network/Tree}
\usage{
AntibodyForest_network_to_pnp_MV(
  AntibodyForest.networks,
  metric.function,
  n.per.sample,
  n.per.network,
  order.networks.per.sample.by.expansion,
  filter.isotype,
  filter.VJcgene,
  species,
  manual_IgKC,
  manual_2A,
  manual_VDJLeader,
  write.to.disk,
  filename,
  whichgroup,
  platypus.version,
  removeintron,
  intron
)
}
\arguments{
\item{AntibodyForest.networks}{List of samples (sample_ids need to be unique), which is a list of clonotypes/igraph objects (AbForest output containing mixcr alignemnt results in network/igraph objects)}

\item{metric.function}{String. Name of vertex attribute to be selected by. It will be selected for nodes having either "yes", TRUE or the highest numeric. (function that can describe metric better (i.e. should be applyable to network and return T or F if it should be included; later on))}

\item{n.per.sample}{Integer or Vector of Integers with length of nbr of samples. Defaults to 1.}

\item{n.per.network}{Integer. Defaults to 1. Enter this value only if metric.function (e.g column name in iGraph function: cell_number == expansion) is a number; will select top n highest values. How many of top expanded clones from each network should be at most picked. E.g., if set to 2, for the top 2 expanded clones PnP sequence constructs will be extracted.}

\item{filter.isotype}{not implemented.}

\item{filter.VJcgene}{not implemented.}

\item{species}{Character. Which IgKC sequence to use. Can be "human" or "mouse". Defaults to "mouse".}

\item{manual_IgKC}{Character. Manual overwrite for sequence used as IgKC.}

\item{manual_2A}{Character. Manual overwrite for sequence used as Furine 2A site.}

\item{manual_VDJLeader}{Character. Manual overwrite for sequence used as VDJ Leader and signal peptide.}

\item{write.to.disk}{Boolean. Defaults to TRUE. Whether to save assembled sequences to working directory.}

\item{filename}{Character. Output file name for .fasta and .csv files if write.to.disk == T. Defaults to PnP_assembled_seqs.fasta/.csv.}

\item{platypus.version}{This function works with "v3" only, there is no need to set this parameter.}
}
\value{
A data.frame containing PnP sequences of selected sequences of networks.
}
\description{
From Networks to PnP Sequences
Extracts from Abforest network function. Initial Version; extracts n most expanded Antibody Sequences per Network from Network/Tree
}
\examples{
\dontrun{
To return results for a non default column
physiologically_relevant <- to_something(VDJ = VDJ_GEX_matrix.output[[1]]
, column.to.plot = "VJ_jgene", normalization factor = 20)
}
}
