% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntibodyForest_network_to_pnp_DL.R
\name{AntibodyForest_node_filter}
\alias{AntibodyForest_node_filter}
\title{Filter function for the AntibodyForest_network_to_pnp function
Input: Dataframe with node features for 1 specific clonotype of 1 specific sample and a (list of) expression(s), which are conditions one wants to filter the clonotypes/networks which are in the ABForest.}
\usage{
AntibodyForest_node_filter(nodes, custom.node.filter)
}
\arguments{
\item{nodes}{Dataframe of node_features extracted from the igraph objects from the AntibodyForest object}

\item{custom.node.filter}{List of expressions, which are to evaluated using eval()}

\item{node.selection}{Boolean vector. Is iterated upon over the conditions specified in custom.node.filter and is returned at the end.}

\item{platypus.version}{This function works with "v3" only, there is no need to set this parameter.}
}
\value{
A boolean vector equaling the number of nodes - which nodes pass the filter
}
\description{
Filter function for the AntibodyForest_network_to_pnp function
Input: Dataframe with node features for 1 specific clonotype of 1 specific sample and a (list of) expression(s), which are conditions one wants to filter the clonotypes/networks which are in the ABForest.
}
