% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntibodyForest_network_to_pnp_DL.R
\name{AntibodyForest_VJcgene_filter}
\alias{AntibodyForest_VJcgene_filter}
\title{Filter function for the AntibodyForest_network_to_pnp function
Input: Dataframe with node features for 1 specific clonotype of 1 specific sample and a vector of character strings, which represent the $VDJ_cgene features one wants to filter}
\usage{
AntibodyForest_VJcgene_filter(
  node_dfs_single,
  i,
  filter.VJcgene,
  established_lc_designation,
  filter.VJcgene.flag
)
}
\arguments{
\item{filter.VJcgene}{Vector of character strings. Should be values found in the $VJ_cgene or $VJchain column}

\item{established_lc_designation}{Character string of established light chain designations. If the strings in filter.VJcgene deviate from the established_lc_designations, it will give a warning.}

\item{node_df_single}{Dataframe of node_features extracted from the igraph objects from the AntibodyForest object - already eventually processed by custom.node.filter}
}
\value{
A boolean vector equaling the number of nodes - which nodes pass the filter
}
\description{
Filter function for the AntibodyForest_network_to_pnp function
Input: Dataframe with node features for 1 specific clonotype of 1 specific sample and a vector of character strings, which represent the $VDJ_cgene features one wants to filter
}
