% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntibodyForests_class.R
\docType{class}
\name{AntibodyForests-class}
\alias{AntibodyForests-class}
\title{Class used for AntibodyForests functions}
\description{
See AntibodyForests function for complete documentation
}
\examples{
\dontrun{
AntibodyForests(VDJ, sequence.type='VDJ.VJ.nt.trimmed',
include.germline=T, network.algorithm='tree',
resolve.ties=c('close.germline.distance', 'max.expansion'),
node.features='OVA_binder', expand.intermediates=T, network.level='intraclonal')
}
}
