% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntibodyForests_degrees.R
\name{AntibodyForests_degrees}
\alias{AntibodyForests_degrees}
\title{Node degree bar plots colored by sequence-level features}
\usage{
AntibodyForests_degrees(
  trees,
  features,
  graph.type,
  which.bipartite,
  max.nodes,
  nbins,
  exclude.germline,
  exclude.intermediates,
  parallel,
  plot.type
)
}
\arguments{
\item{trees}{AntibodyForests object/list of AntibodyForests objects - the resulting sequence similarity or minimum spanning tree networks from the AntibodyForests function}

\item{features}{vector of strings - sequence-level features for the bar plot colors.}

\item{graph.type}{string - the graph type available in the AntibodyForests object which will be used as the function input.
Currently supported network/analysis types: 'tree' (for the minimum spanning trees or sequence similarity networks obtained from the main AntibodyForests function), 'heterogeneous' for the bipartite graphs obtained via AntibodyForests_heterogeneous.}

\item{which.bipartite}{string - whether to create a bar plot for the cell layer of the bipartite/heterogeneous graph ('cells'), sequence layer ('sequences') or for both ('both').}

\item{max.nodes}{integer - the maximum number of nodes (bars) to be plotted in the degree bar plot}

\item{nbins}{integer - number of bins for the degree histograms if plot.type == 'histogram'}

\item{exclude.germline}{boolean - if T, will exclude the germline nodes when creating the bar plots.}

\item{exclude.intermediates}{boolean - if T, will exclude the intermediate nodes when creating the bar plots (if expand.intermediates was set to T when creating the AntibodyForests object).}

\item{parallel}{boolean - if T, will execute the main loop in parallel using mclapply.}

\item{plot.type}{string - 'barplot' for ordered node degree bar plots colored by the sequence-level feature specified in the features parameter. 'histogram' for node degree histograms.}
}
\value{
a ggplot2 bar plot or histogram.
}
\description{
Creates an ordered bar plot of the vertices with the highest degree, colored by the sequence-level features specified in the features parameter. Creates a bar plot per sample, per clonotype, or per entire AntibodyForests object, depending on the input format (object, list, nested list).
Secondary option to create a node degree histogram if plot.type == 'histogram'.
}
\examples{
\dontrun{
AntibodyForests_object |> AntibodyForests_degrees(features = 'VDJ_cgene', plot.type = 'barplot')
}
}
\seealso{
AntibodyForests, AntibodyForests_metrics
}
