% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_VDJ_iteration.R
\name{barcode_VDJ_iteration}
\alias{barcode_VDJ_iteration}
\title{Helper function called in VDJ_GEX_matrix. Do not run as standalone!}
\usage{
barcode_VDJ_iteration(
  barcodes,
  contigs,
  references,
  annotations,
  gap.opening.cost,
  gap.extension.cost,
  trim.and.align
)
}
\arguments{
\item{All_PARAMS}{Inputs provided by VDJ_GEX_matrix}
}
\value{
List of tables, with each table containing information one one barcode
}
\description{
Helper function called in VDJ_GEX_matrix. Do not run as standalone!
}
\examples{
\dontrun{
Do not run as standalone!
}
}
