% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbForests_data.R
\docType{data}
\name{Bcell_sequences_example_tree}
\alias{Bcell_sequences_example_tree}
\title{Example csv file 1}
\format{
An object of class \code{data.frame} with 170 rows and 1 columns.
}
\usage{
Bcell_sequences_example_tree
}
\description{
Example csv file 1
}
\references{
R package Platypus : https://doi.org/10.1093/nargab/lqab023
}
\keyword{datasets}
