% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbForests_data.R
\docType{data}
\name{Bcell_tree_2}
\alias{Bcell_tree_2}
\title{Example csv file 2}
\format{
An object of class \code{data.frame} with 85 rows and 1 columns.
}
\usage{
Bcell_tree_2
}
\description{
Example csv file 2
}
\references{
R package Platypus:https://doi.org/10.1093/nargab/lqab023
}
\keyword{datasets}
