% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_net.R
\docType{data}
\name{binary_colormap}
\alias{binary_colormap}
\title{Binary segmentation color map.}
\format{
An object of class \code{list} of length 2.
}
\usage{
binary_colormap
}
\value{
Binary segmentation color map.
}
\description{
Binary segmentation color map.
}
\keyword{datasets}
