% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_net.R
\docType{data}
\name{binary_labels}
\alias{binary_labels}
\title{Binary segmentation labels.}
\format{
An object of class \code{character} of length 2.
}
\usage{
binary_labels
}
\value{
Binary segmentation labels.
}
\description{
Binary segmentation labels.
}
\keyword{datasets}
