% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_anchors.R
\name{box_jaccard_distance}
\alias{box_jaccard_distance}
\title{Calculates Jaccard distance between two boxes.}
\usage{
box_jaccard_distance(box1_w, box1_h, box2_w, box2_h)
}
\arguments{
\item{box1_w}{Width of first box.}

\item{box1_h}{Height of first box.}

\item{box2_w}{Width of second box.}

\item{box2_h}{Height of second box.}
}
\value{
Jaccard distance between two boxes.
}
\description{
Calculates Jaccard distance between two boxes.
}
