% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_loss.R
\name{calculate_iou}
\alias{calculate_iou}
\title{Calculates boxes IoU.}
\usage{
calculate_iou(pred_boxes, true_boxes)
}
\arguments{
\item{pred_boxes}{Tensor of predicted coordinates.}

\item{true_boxes}{Tensor of true coordinates.}
}
\value{
IoU between true and predicted boxes.
}
\description{
Calculates boxes IoU.
}
