% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_predict.R
\name{check_boxes_intersect}
\alias{check_boxes_intersect}
\title{Checks if two bounding boxes intersect.}
\usage{
check_boxes_intersect(box1, box2)
}
\arguments{
\item{box1}{Vector `(xmin, ymin, xmax, ymax)` with box coordinates.}

\item{box2}{Vector `(xmin, ymin, xmax, ymax)` with box coordinates.}
}
\value{
`TRUE` if `box1` and `box2` intersect, else `FALSE`.
}
\description{
Checks if two bounding boxes intersect.
}
