% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Echidna_data.R
\docType{data}
\name{class_switch_prob_hum}
\alias{class_switch_prob_hum}
\title{class_switch_prob_hum
The probability matrix of class switching for human b cells. The row names of the matrix are the isotypes the cell is switching from, the column names are the isotypes the cell is switching to. All B cells start from IGHM, and switch to one of the other isotypes or remain the same.}
\format{
A 8*8 matrix.The row and clumn names are "IGHM","IGHD","IGHG1","IGHG2","IGHG3","IGHG4","IGHE","IGHA".
The probability for a cell to switch from "IGHM" to "IGHD" is the value at class_switch_prob_hum[1,2].
}
\usage{
data("class_switch_prob_hum")
}
\description{
class_switch_prob_hum
The probability matrix of class switching for human b cells. The row names of the matrix are the isotypes the cell is switching from, the column names are the isotypes the cell is switching to. All B cells start from IGHM, and switch to one of the other isotypes or remain the same.
}
\keyword{datasets}
