% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_predict.R
\name{clean_boxes}
\alias{clean_boxes}
\title{Transform list of bounding boxes into a `data.frame`.}
\usage{
clean_boxes(boxes, labels)
}
\arguments{
\item{boxes}{List of bounding boxes.}

\item{labels}{Character vector containing class labels. For example \code{\link[platypus]{coco_labels}}.}
}
\value{
List of `data.frames` containing bounding box coordinates and objectness/class scores.
}
\description{
Transform list of bounding boxes into a `data.frame`.
}
