% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Echidna_other_functions.R
\name{clonofreq.isotype.data}
\alias{clonofreq.isotype.data}
\title{Get information about the clonotype counts grouped by isotype.}
\usage{
clonofreq.isotype.data(all.contig.annotations, top.n)
}
\arguments{
\item{all.contig.annotations}{The output dataframe all_contig_annotation from function simulate.repertoire.}

\item{top.n}{The top n abundant clones to be shown in the plot. If missing, all clones will be shown.}
}
\value{
dataframes containing the top n abundant clonotypes and their frequency and isotype information for further processing.
}
\description{
Return
}
