% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Echidna_other_functions.R
\name{clonofreq.isotype.plot}
\alias{clonofreq.isotype.plot}
\title{Get information about the clonotype counts grouped by isotype.}
\usage{
clonofreq.isotype.plot(all.contig.annotations, top.n, y.limit, colors)
}
\arguments{
\item{all.contig.annotations}{The output dataframe all_contig_annotation from function simulate.repertoire.}

\item{top.n}{The top n abundant clones to be shown in the plot. If missing, all clones will be shown.}

\item{y.limit}{The upper limit for y axis in the plot.}

\item{colors}{A named character vector of colors, the names are the isotypes. If missing, the default has 11 colors coresponding to the default isotype names.}
}
\value{
a stacked barplot for clonotype counts grouped by isotype
}
\description{
Plot a stacked barplot for clonotype counts grouped by isotype.
}
