% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Echidna_other_functions.R
\name{clonofreq}
\alias{clonofreq}
\title{Plot clonal frequency barplot of the outout simulated data}
\usage{
clonofreq(clonotypes, top.n, y.limit)
}
\arguments{
\item{clonotypes}{The clonotypes dataframe, which is the second element in the simulation output list.}

\item{top.n}{The top n abundant clones to be shown in the plot. If missing, all clones will be shown.}

\item{y.limit}{The upper limit for y axis in the plot.}
}
\value{
top abundant clonal frequencies in a barplot with ggplot2
}
\description{
Plot the top abundant clonal frequencies in a barplot with ggplot2
}
