% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Echidna_other_functions.R
\name{clonofreq.trans.data}
\alias{clonofreq.trans.data}
\title{Get information about the clonotype counts grouped by transcriptome state(cell type).}
\usage{
clonofreq.trans.data(all.contig.annotations, history, trans.names, top.n)
}
\arguments{
\item{all.contig.annotations}{The output dataframe all_contig_annotation from function simulate.repertoire.}

\item{history}{The dataframe history from simulate output.}

\item{trans.names}{The names of cell types which are used in transcriptome.switch.prob argument in the simulation.}

\item{top.n}{The top n abundant clones to be shown in the plot. If missing, all clones will be shown.}
}
\value{
a dataframe with clonotype counts grouped by transcriptome state(cell type).
}
\description{
Dataframe with clonotype counts grouped by transcriptome state(cell type).
}
