% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Echidna_other_functions.R
\name{clonofreq.trans.plot}
\alias{clonofreq.trans.plot}
\title{Get information about the clonotype counts grouped by transcriptome state(cell type).}
\usage{
clonofreq.trans.plot(
  all.contig.annotations,
  history,
  trans.names,
  top.n,
  y.limit,
  colors
)
}
\arguments{
\item{all.contig.annotations}{The output dataframe all_contig_annotation from function simulate.repertoire.}

\item{history}{The dataframe history from simulate output.}

\item{trans.names}{The names of cell types which are used in transcriptome.switch.prob argument in the simulation.}

\item{top.n}{The top n abundant clones to be shown in the plot. If missing, all clones will be shown.}

\item{y.limit}{The upper limit for y axis in the plot.}

\item{colors}{A named character vector of colors, the names are the isotypes. If missing, the default has 11 colors coresponding to the default isotype names.}
}
\value{
a stacked barplot for clonotype counts grouped by transcriptome state(cell type).
}
\description{
Plot a stacked barplot for clonotype counts grouped by transcriptome state(cell type).
}
