% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Echidna_other_functions.R
\name{cluster.id.igraph}
\alias{cluster.id.igraph}
\title{Get clone network igraphs colored by seurat cluster id.}
\usage{
cluster.id.igraph(meta.data, history, igraph.index, empty.node)
}
\arguments{
\item{meta.data}{the meta.data dataframe from the Seurat object of the simulation. The object should be pre-processed and has cluster ids in the meta.data.}

\item{history}{The dataframe 'history' from the simulation output.}

\item{igraph.index}{The list 'igraph.index' from the simulation output.}

\item{empty.node}{If TRUE, there will be empty node in igraph. if FALSE, the empty node will be deleted.}
}
\description{
Get clone network igraphs colored by seurat cluster id.
}
