% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3.R
\docType{data}
\name{coco_anchors}
\alias{coco_anchors}
\title{COCO dataset anchors.}
\format{
An object of class \code{list} of length 3.
}
\usage{
coco_anchors
}
\value{
COCO dataset anchors.
}
\description{
COCO dataset anchors.
}
\keyword{datasets}
