% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3.R
\docType{data}
\name{coco_labels}
\alias{coco_labels}
\title{COCO dataset labels.}
\format{
An object of class \code{character} of length 80.
}
\usage{
coco_labels
}
\value{
COCO dataset labels.
}
\description{
COCO dataset labels.
}
\keyword{datasets}
