% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_predict.R
\name{correct_boxes}
\alias{correct_boxes}
\title{Rescales boxes.}
\usage{
correct_boxes(boxes, image_h, image_w)
}
\arguments{
\item{image_h}{Rescaling factor for `ymin/ymax` box coordinates.}

\item{image_w}{Rescaling factor for `xmin/xmax` box coordinates.}
}
\value{
Rescaled bounding boxes.
}
\description{
Rescales boxes. `xmin/xmax` coordinates are multiplied by `image_w` and `ymin/ymax` coordinates are multiplied by `image_h`.
}
