% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_generator.R
\name{create_images_masks_paths}
\alias{create_images_masks_paths}
\title{Generates inmages/masks path from selected cofiguration.}
\usage{
create_images_masks_paths(path, mode, only_images, subdirs = c("images",
  "masks"), column_sep = ";")
}
\arguments{
\item{path}{Images directories or configuration file path.}

\item{mode}{Character. One of `c("dir", "nested_dirs", "config_file")` or `c(1, 2, 3)` correspondingly.}

\item{only_images}{Should generator read only images (e.g. on train set for predictions).}

\item{subdirs}{Vector of two characters containing names of subdirectories with images and masks.}

\item{column_sep}{Character. Configuration file separator.}
}
\description{
Generates inmages/masks path from selected cofiguration.
}
