% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_segmentation_map_ggplot}
\alias{create_segmentation_map_ggplot}
\title{Generates raster image with segmentation masks.}
\usage{
create_segmentation_map_ggplot(
  image_path,
  masks,
  labels,
  colormap,
  target_size,
  grayscale
)
}
\arguments{
\item{image_path}{Image filepath.}

\item{masks}{Predicted segmentation masks.}

\item{labels}{Character vector containing class labels. For example \code{\link[platypus]{voc_labels}}.}

\item{colormap}{Class color map. For example \code{\link[platypus]{voc_colormap}}.}

\item{target_size}{Image target size.}

\item{grayscale}{Should images be plotted in grayscale.}
}
\value{
Raster image with segmentation masks.
}
\description{
Generates raster image with segmentation masks.
}
