% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_train_on_batch.R
\name{custom_predict_generator}
\alias{custom_predict_generator}
\title{Calculates predictions on new samples using data generator.}
\usage{
custom_predict_generator(model, generator, steps)
}
\arguments{
\item{model}{Model.}

\item{generator}{Data generator.}

\item{steps}{Steps in epoch.}
}
\description{
Calculates predictions on new samples using data generator.
}
