% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.platypus.R
\name{cv.platypus}
\alias{cv.platypus}
\title{\code{Pseudocode:
    get platypus parameters and view information
    load labelled data
    divide labelled data in k subsets
    for fold.i in 1:k
      hold out subset fold.i of labelled data
      train platypus.R on all-but-fold.i labelled (and unlabelled) data with the given views   # unlabelled data is optional, but making use of unlabelled data is one of the biggest advantages of platypus
      predict fold.i subset on trained platypus views
}
k-fold cross validation for platypus}
\usage{
cv.platypus(fn.views, fn.labs, classcol.labs = 1, cv.folds = 10,
  no.iterations = 100, majority.threshold.percent = 100,
  expanded.output = FALSE, updating = FALSE,
  ignore.label = "intermediate", parallel = FALSE, num.cores = 25,
  output.folder = NA)
}
\arguments{
\item{fn.views}{List of view files}

\item{fn.labs}{File containing outcome labels}

\item{k}{number of folds for label learning validation (similar to cross validation folds), eg. 5, default=10}

\item{i}{Maximal number of iterations for each platypus run, default=100}

\item{m}{Percent agreement required to learn a sample's class label, default=100}

\item{u}{Updating the accuracies of the single views in each iteration, default=FALSE}

\item{e}{Expanded output: returned result list contains a list of trained views after each iteration, default=FALSE}

\item{b}{Label class to ignore, if any. Defaults to 'intermediate'}

\item{o}{Name of the folder where output is stored.}

\item{p}{The number of cores to use. Enables parallelization.}
}
\value{
A list containing fold.accuracy, labelling.matrix,labelling.matrices.views
}
\description{
\code{Pseudocode:
    get platypus parameters and view information
    load labelled data
    divide labelled data in k subsets
    for fold.i in 1:k
      hold out subset fold.i of labelled data
      train platypus.R on all-but-fold.i labelled (and unlabelled) data with the given views   # unlabelled data is optional, but making use of unlabelled data is one of the biggest advantages of platypus
      predict fold.i subset on trained platypus views
}
k-fold cross validation for platypus
}
\examples{
TODO show how to generate config.files and fn.labs
cv.platypus(fn.views=config.files,fn.labs=fn.labs,no.iterations=5,majority.threshold.percent=75,output.folder='platypus_output')
cv.platypus(config.files,fn.labs)
}
\keyword{platypus}
