% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darknet53.R
\name{darknet53}
\alias{darknet53}
\title{Creates a Darknet53 architecture.}
\usage{
darknet53(channels)
}
\arguments{
\item{channels}{Number of channels.}
}
\value{
Darknet53 model.
}
\description{
Creates a Darknet53 architecture.
}
