% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darknet53.R
\name{darknet53_conv2d}
\alias{darknet53_conv2d}
\title{Creates a convolutional Darknet53 unit.}
\usage{
darknet53_conv2d(
  input,
  strides,
  filters,
  kernel_size,
  batch_normalization = TRUE,
  leaky_relu = TRUE
)
}
\arguments{
\item{input}{Model or layer object.}

\item{strides}{An integer or list of 2 integers, specifying the strides of the convolution along the width and height.}

\item{filters}{Integer, the dimensionality of the output space (i.e. the number of output filters in the convolution).}

\item{kernel_size}{An integer or list of 2 integers, specifying the width and height of the 2D convolution window. Can be a single integer to specify the same value for all spatial dimensions.}

\item{batch_normalization}{Should batch normalization be used in the unit.}

\item{leaky_relu}{Should leaky ReLU activation function be used in the unit.}
}
\value{
Convolutional Darknet53 unit.
}
\description{
Creates a convolutional Darknet53 unit.
}
