% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darknet53.R
\name{darknet53_residual_block}
\alias{darknet53_residual_block}
\title{Creates a residual Darknet53 block.}
\usage{
darknet53_residual_block(input, filters, blocks)
}
\arguments{
\item{input}{Model or layer object.}

\item{filters}{Integer, the dimensionality of the output space (i.e. the number of output filters in the convolution).}

\item{blocks}{Number of residual blocks.}
}
\value{
Residual Darknet53 block.
}
\description{
Creates a residual Darknet53 block.
}
