% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_loss_metric.R
\name{dice_coef}
\alias{dice_coef}
\title{Function for calculating dice coefficient.}
\usage{
dice_coef(y_true, y_pred, smooth = 1)
}
\arguments{
\item{y_true}{Tensor of true probabilities.}

\item{y_pred}{Tensor of predicted probabilities.}

\item{smooth}{Smoothing Factor.}
}
\value{
Dice coefficient.
}
\description{
Function for calculating dice coefficient.
}
