% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_generator.R
\name{find_anchors_and_grids_for_true_boxes}
\alias{find_anchors_and_grids_for_true_boxes}
\title{Finds best anchors and output grid coordinates for true bounding box coordinates.}
\usage{
find_anchors_and_grids_for_true_boxes(true_boxes, anchors, true_grid)
}
\arguments{
\item{true_grid}{`Yolo3` output grids.}

\item{batch_size}{Batch size.}

\item{net_h}{Input layer height from trained \code{\link[platypus]{yolo3}} model. Must be divisible by `32`.}

\item{net_w}{Input layer width from trained \code{\link[platypus]{yolo3}} model. Must be divisible by `32`.}

\item{downscale}{Integer specifying how to downscale `net_h` and `net_w`. For `Yolo3` output grids equal do `32`, `16` and `8`.}
}
\value{
`data.frame` with best anchors and output grid coordinates for true bounding box coordinates.
}
\description{
Finds best anchors and output grid coordinates for true bounding box coordinates.
}
