% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.config.R
\name{gen.config}
\alias{gen.config}
\title{Generate configuration files for platypus}
\usage{
gen.config(view.names, fn.tasks, config.loc = "config",
  model.type = c("en", "rf", "svm"), delim = ",", delim.v = "\\t",
  n.iters = 10, ignore.label = "intermediate", nfolds = 10,
  mtry = NA, ntree = c(500, 1000, 1500, 2000))
}
\arguments{
\item{view.names}{List of files containing view feature data}

\item{fn.tasks}{File containing all task labels, one column per task}

\item{config.loc}{Where the config files should be stored}

\item{model.type}{Type of classifier to use}
}
\value{
List of config filenames, for use in platypus
}
\description{
Generate configuration files for platypus
}
