% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_anchors.R
\name{generate_anchors}
\alias{generate_anchors}
\title{Calculates anchor boxes using k-mean++ algorithm.}
\usage{
generate_anchors(
  anchors_per_grid,
  annot_paths,
  labels,
  net_h,
  net_w,
  n_iter = 10,
  seed = 1234,
  centroid_fun = mean
)
}
\arguments{
\item{anchors_per_grid}{Number of anchors per one grid.}

\item{annot_paths}{List of annotations filepaths.}

\item{labels}{Character vector containing class labels. For example \code{\link[platypus]{coco_labels}}.}

\item{net_h}{Input layer height from trained \code{\link[platypus]{yolo3}} model. Must be divisible by `32`.}

\item{net_w}{Input layer width from trained \code{\link[platypus]{yolo3}} model. Must be divisible by `32`.}

\item{n_iter}{Maximum number of iteration for k-mean++ algorithm.}

\item{seed}{Random seed.}

\item{centroid_fun}{Function to use for centroid calculation.}
}
\value{
List of anchor boxes.
}
\description{
Calculates anchor boxes using k-mean++ algorithm.
}
