% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_generator.R
\name{generate_empty_grid}
\alias{generate_empty_grid}
\title{Creates empty output `Yolo3` grid.}
\usage{
generate_empty_grid(
  batch_size,
  net_h,
  net_w,
  downscale,
  anchors_per_grid,
  n_class
)
}
\arguments{
\item{batch_size}{Batch size.}

\item{net_h}{Input layer height from trained \code{\link[platypus]{yolo3}} model. Must be divisible by `32`.}

\item{net_w}{Input layer width from trained \code{\link[platypus]{yolo3}} model. Must be divisible by `32`.}

\item{downscale}{Integer specifying how to downscale `net_h` and `net_w`. For `Yolo3` output grids equal do `32`, `16` and `8`.}
}
\value{
Empty output `Yolo3` grid.
}
\description{
Creates empty output `Yolo3` grid.
}
