% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Echidna_other_functions.R
\name{get.avr.mut.data}
\alias{get.avr.mut.data}
\title{Get information about somatic hypermutation in the simulation. This function return a barplot showing the average mutation.}
\usage{
get.avr.mut.data(igraph.index.attr, history, clonotype.select, level)
}
\arguments{
\item{igraph.index.attr}{A list "igraph.index.attr" from the simulation output.}

\item{history}{A dataframe "history" from the simulation output.}

\item{clonotype.select}{The selected clonotype index, can be the output of the function "select.top".}

\item{level}{Can be "clone" or "cell". If "clone", the function will return average mutation on unique variant level. Otherwise it will return on cell level.}
}
\description{
Get information about somatic hypermutation in the simulation. This function return a barplot showing the average mutation.
}
