% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Echidna_other_functions.R
\name{get.avr.mut.plot}
\alias{get.avr.mut.plot}
\title{Get information about somatic hypermutation in the simulation. This function return a barplot showing the average mutation.}
\usage{
get.avr.mut.plot(igraph.index.attr, history, clonotype.select, level, y.limit)
}
\arguments{
\item{igraph.index.attr}{A list "igraph.index.attr" from the simulation output.}

\item{history}{A dataframe "history" from the simulation output.}

\item{clonotype.select}{The selected clonotype index, can be the output of the function "select.top".}

\item{level}{Can be "node" or "cell". If "node", the function will return average mutation on unique variant level. Otherwise it will return on cell level.}

\item{y.limit}{The upper limit for y axis in the plot.}
}
\value{
a barplot showing the average mutation per node (same heavy and light chain set) or per cell.
}
\description{
Get information about somatic hypermutation in the simulation. This function return a barplot showing the average mutation.
}
