% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Echidna_other_functions.R
\name{get.barplot.errorbar}
\alias{get.barplot.errorbar}
\title{Return a barplot of mean and standard error bar of certain value of each clone.}
\usage{
get.barplot.errorbar(data, y.lab, y.limit)
}
\arguments{
\item{data}{A dataframe. Columns are different simulations, rows are the top clones. The first row is the top abundant clone.}

\item{y.lab}{A string specifies the y lable name of the barplot.}

\item{y.limit}{The upper limit for y axis in the plot.}
}
\value{
a barplot of mean and standard error bar of certain value of each clone.
}
\description{
Return a barplot of mean and standard error bar of certain value of each clone.
}
