% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Echidna_other_functions.R
\name{get.elbow}
\alias{get.elbow}
\title{Get the seurat object from simulated transciptome output.}
\usage{
get.elbow(data)
}
\arguments{
\item{data}{The output "transcriptome" dataframe from simulation output.}
}
\description{
Get the seurat object from simulated transciptome output.
}
