% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Echidna_other_functions.R
\name{get.n.node.data}
\alias{get.n.node.data}
\title{Get the number of unique variants in each clone in a vector. The output is the vector representing the numbers of unique variants.}
\usage{
get.n.node.data(data, clonotype.select)
}
\arguments{
\item{data}{The output "igraph.index.attr" list from simulation output.}

\item{clonotype.select}{The index of the clones to be shown. If missing, all clones will be included.}
}
\description{
Get the number of unique variants in each clone in a vector. The output is the vector representing the numbers of unique variants.
}
