% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Echidna_other_functions.R
\name{get.n.node.plot}
\alias{get.n.node.plot}
\title{Get the number of unique variants in each clone in a vector and the barplot. The first item in the output is the vector representing the numbers of unique variants, the second item is the barplot.}
\usage{
get.n.node.plot(igraph.index.attr, clonotype.select, y.limit)
}
\arguments{
\item{igraph.index.attr}{The output "igraph.index.attr" list from simulation output.}

\item{clonotype.select}{The index of the clones to be shown. If missing, all clones will be included.}

\item{y.limit}{The upper limit for y axis in the plot.}
}
\value{
the number of unique variants in each clone in a vector and the barplot. The first item in the output is the vector representing the numbers of unique variants, the second item is the barplot.
}
\description{
Get the number of unique variants in each clone in a vector and the barplot. The first item in the output is the vector representing the numbers of unique variants, the second item is the barplot.
}
