% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Echidna_other_functions.R
\name{get.seq.distance}
\alias{get.seq.distance}
\title{Computing sequence distance according to the number of unmatched bases.}
\usage{
get.seq.distance(germline, sequence)
}
\arguments{
\item{germline}{A string representing the germline sequence.}

\item{sequence}{A string of the sequence to be compared, which has the same length as germline.}
}
\value{
the number of unmatched bases in 2 sequences.
}
\description{
Computing sequence distance according to the number of unmatched bases.
}
