% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Echidna_other_functions.R
\name{get.vgu.matrix}
\alias{get.vgu.matrix}
\title{Get paired v gene heavy chain and light chain matrix on clonotype level. A v gene usage pheatmap can be obtain by p<-pheatmap::pheatmap(vgu_matrix,show_colnames= T, main = "V Gene Usage"), where the vgu_matrix is the output of this function.}
\usage{
get.vgu.matrix(all.contig.annotations, level)
}
\arguments{
\item{all.contig.annotations}{The dataframe "all_contig_annotation" from simulation output.}

\item{level}{Can be "clone" or "cell". If "clone", the function will return paired v gene usage matrix on clonotype level. Otherwise it will return on cell level.}
}
\value{
a paired v gene heavy chain and light chain matrix on clonotype level.
}
\description{
Get paired v gene heavy chain and light chain matrix on clonotype level. A v gene usage pheatmap can be obtain by p<-pheatmap::pheatmap(vgu_matrix,show_colnames= T, main = "V Gene Usage"), where the vgu_matrix is the output of this function.
}
