% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_predict.R
\name{get_boxes}
\alias{get_boxes}
\title{Transforms Yolo3 predictions into valid boxes.}
\usage{
get_boxes(
  preds,
  anchors,
  labels,
  obj_threshold = 0.6,
  net_h = 416,
  net_w = 416,
  nms = TRUE,
  nms_threshold = 0.6,
  correct_hw = FALSE,
  image_h = NULL,
  image_w = NULL
)
}
\arguments{
\item{preds}{\code{\link[platypus]{yolo3}} model predictions.}

\item{anchors}{Prediction anchors. For exact format check \code{\link[platypus]{coco_anchors}}.}

\item{labels}{Character vector containing class labels. For example \code{\link[platypus]{coco_labels}}.}

\item{obj_threshold}{Minimum objectness score. Must be in range `[0, 1]`. All boxes with objectness score less than `obj_threshold` will be filtered out.}

\item{net_h}{Input layer height from trained \code{\link[platypus]{yolo3}} model. Must be divisible by `32`.}

\item{net_w}{Input layer width from trained \code{\link[platypus]{yolo3}} model. Must be divisible by `32`.}

\item{nms}{Logical. Should `Non-Maximum-Suppression` be applied.}

\item{nms_threshold}{`Non-Maximum-Suppression` threshold.}

\item{correct_hw}{Logical. Should height/width rescaling of bounding boxes be applied. If `TRUE` `xmin/xmax` coordinates are multiplied by `image_w` and `ymin/ymax` coordinates are multiplied by `image_h`.}

\item{image_h}{Rescaling factor for `ymin/ymax` box coordinates.}

\item{image_w}{Rescaling factor for `xmin/xmax` box coordinates.}
}
\value{
List of `data.frames` containing bounding box coordinates and objectness/class scores.
}
\description{
Transforms Yolo3 predictions into valid boxes.
}
