% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_predict.R
\name{get_masks}
\alias{get_masks}
\title{Transforms `U-Net` predictions into valid segmentation map.}
\usage{
get_masks(preds, colormap)
}
\arguments{
\item{preds}{\code{\link[platypus]{u_net}} model predictions.}

\item{colormap}{Class color map. For example \code{\link[platypus]{voc_colormap}}.}
}
\description{
Transforms `U-Net` predictions into valid segmentation map.
}
