% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_loss.R
\name{get_max_boxes_iou}
\alias{get_max_boxes_iou}
\title{Compares boxes by IoU.}
\usage{
get_max_boxes_iou(pred_boxes, true_boxes)
}
\arguments{
\item{pred_boxes}{Tensor of predicted coordinates.}

\item{true_boxes}{Tensor of true coordinates.}
}
\value{
Max IoU between true and predicted boxes.
}
\description{
Compares boxes by IoU.
}
