% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_generator.R
\name{get_true_boxes_from_annotations}
\alias{get_true_boxes_from_annotations}
\title{Calculates true bounding box coordinates from annotations.}
\usage{
get_true_boxes_from_annotations(
  annotations,
  net_h,
  net_w,
  anchors,
  labels,
  true_grid
)
}
\arguments{
\item{annotations}{Annotations.}

\item{net_h}{Input layer height from trained \code{\link[platypus]{yolo3}} model. Must be divisible by `32`.}

\item{net_w}{Input layer width from trained \code{\link[platypus]{yolo3}} model. Must be divisible by `32`.}

\item{anchors}{Prediction anchors. For exact format check \code{\link[platypus]{coco_anchors}}.}

\item{labels}{Character vector containing class labels. For example \code{\link[platypus]{coco_labels}}.}

\item{true_grid}{`Yolo3` output grids.}
}
\value{
`data.frame` with best anchors and output grid coordinates for true bounding box coordinates.
}
\description{
Calculates true bounding box coordinates from annotations.
}
