% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEX_automate_single.R
\name{GEX_automate_single}
\alias{GEX_automate_single}
\title{Helper function for GEX processing in VDJ_GEX_matrix. DO NOT RUN AS STANDALONE}
\usage{
GEX_automate_single(
  GEX.list,
  GEX.integrate,
  integration.method,
  VDJ.gene.filter,
  mito.filter,
  norm.scale.factor,
  n.feature.rna,
  n.count.rna.min,
  n.count.rna.max,
  n.variable.features,
  cluster.resolution,
  neighbor.dim,
  mds.dim,
  group.id
)
}
\arguments{
\item{ALL_PARAMS}{Provided by VDJ_GEX_matrix function.}
}
\value{
GEX matrix
}
\description{
Helper function for GEX processing in VDJ_GEX_matrix. DO NOT RUN AS STANDALONE
}
\examples{
\dontrun{
Do not run as standalone!
}
}
