% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEX_cluster_genes_heatmap.R
\name{GEX_cluster_genes_heatmap1}
\alias{GEX_cluster_genes_heatmap1}
\title{Produces a heatmap displaying the expression of the top genes that define each cluster in the Seurat object. The output heatmap is derived from DoHeatmap from Seurat and thereby can be edited using typical ggplot interactions. The number of genes per cluster and the nunber of cells to display can be specified by the user. Either the log fold change or the p value can be used to select the top n genes.}
\usage{
GEX_cluster_genes_heatmap1(
  automate_GEX.output,
  GEX_cluster_genes.output,
  n.genes.per.cluster,
  metric,
  max.cell
)
}
\arguments{
\item{automate_GEX.output}{Output Seurat object containing gene expression data from automate_GEX function that contained at least two distinct biological samples. The different biological samples correspond to integer values in the order of the working directories initially supplied to the automate_GEX function.}

\item{GEX_cluster_genes.output}{The output from the GEX_cluster_genes function - this should be a list with each list element corresponding to the genes, p values, logFC, pct expression for the genes differentially regulated for each cluster.}

\item{n.genes.per.cluster}{An integer value determining how many genes per cluster to display in the output heatmap. This number should be adjusted based on the number of clusters. Too many genes per cluster and clusters may cause a problem with the heatmap function in Seurat.}

\item{metric}{The metric that dictates which are the top n genes returned. Possible options are "p.value" (default), "avg_logFC", "top_logFC", "bottom_logFC". "top_logFC" returns the top expressed genes for each cluster, whereas "bottom_logFC" returns the least expressed genes per cluster-both by log fold change.}

\item{max.cell}{The max number of cells to display in the heatmap for each cluster, which corresponds to the number of columns. Default is set to 100 cells per cluster.}
}
\value{
Returns a heatmap from the function DoHeatmap from the package Seurat, which is a ggplot object that can be modified or plotted. The number of genes is determined by the n.genes parameter and the number of cells per cluster is determined by the max.cell argument. This function gives a visual description of the top genes differentially expressed in each cluster.
}
\description{
Produces a heatmap displaying the expression of the top genes that define each cluster in the Seurat object. The output heatmap is derived from DoHeatmap from Seurat and thereby can be edited using typical ggplot interactions. The number of genes per cluster and the nunber of cells to display can be specified by the user. Either the log fold change or the p value can be used to select the top n genes.
}
\examples{
\dontrun{
cluster_defining_gene_heatmap <- GEX_cluster_genes_heatmap(automate_GEX.output=automate_GEX_output[[i]],GEX_cluster_genes.output=GEX_cluster_genes_output,n.genes.per.cluster=5,metric="p.value",max.cell=5)
}
}
