% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEX_cluster_per_sample.R
\name{GEX_cluster_membership_per_sample}
\alias{GEX_cluster_membership_per_sample}
\title{Plots the cluster membership for each of the distinct samples in the Seurat object from the automate_GEX function. The distinct samples are determined by "sample_id" field in the Seurat object.}
\usage{
GEX_cluster_membership_per_sample(automate_GEX.output, by.group)
}
\arguments{
\item{automate_GEX.output}{Output Seurat object containing gene expression data from automate_GEX function that contained at least two distinct biological samples. The different biological samples correspond to integer values in the order of the working directories initially supplied to the automate_GEX function.}

\item{by.group}{Logical indicating whether to look at the cluster distributin per group. Default is set to FALSE.}
}
\value{
Returns a ggplot in which the values on the x axis correspond to each cluster found in the Seurat object. The y axis corresponds to the percentage of cells found in each cluster. The bar and color corresponds to the distinct sample_id.
}
\description{
Plots the cluster membership for each of the distinct samples in the Seurat object from the automate_GEX function. The distinct samples are determined by "sample_id" field in the Seurat object.
}
\examples{
\dontrun{
cluster.distribution.per.sample <- GEX_cluster_membership_per_sample(automate_GEX.output=automate_GEX_out[[i]])
}
}
