% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEX_fraction_cells_per_cluster.R
\name{GEX_fraction_cells_per_cluster}
\alias{GEX_fraction_cells_per_cluster}
\title{Plots the fraction of cells in each cluster compared between sample or group.}
\usage{
GEX_fraction_cells_per_cluster(GEX.object, per_sample)
}
\arguments{
\item{GEX.object}{Output of the automate_GEX function}

\item{per_sample}{Logical indicating if the grouping is done at the level of the samples or the groups}
}
\description{
Plots the fraction of cells in each cluster compared between sample or group.
}
\examples{
\dontrun{
GEX_fraction_cells_per_cluster(automate_GEX.output, per_sample))
}
}
