% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEX_GSEA.R
\name{GEX_GSEA}
\alias{GEX_GSEA}
\title{Conducts a Gene Set Enrichment Analysis (GSEA) on a set of genes submitted in a data frame with a metric each. Works with the output of GEX_genes_cluster or a custom data frame contianing the gene symbols either in a column "symbols" or as rownames and a metric for each gene. The name of the column containing the metric has to be declared via the input metric_colname.}
\usage{
GEX_GSEA(
  GEX.cluster.genes.output,
  MT.Rb.filter,
  filter,
  path_to_pathways,
  metric_colname,
  pval_adj_cutoff,
  Enrichment.Plots,
  my_own_geneset
)
}
\arguments{
\item{GEX.cluster.genes.output}{Data frame containing the list of gene symbols and a metric. Function works directly with GEX_cluster_genes output.}

\item{MT.Rb.filter}{Logical, should Mitotic and Ribosomal genes be filtered out of the geneset. True by default.}

\item{filter}{Character vector containing the identifying symbol sequence for the genes which should be filtered out, if MT.Rb.filter == T. By default set to c("MT-", "RPL", "RPS").}

\item{path_to_pathways}{Either a path to gmt file containing the gene sets (can be downloaded from MSigDB) or vector where first element specifies species and second element specifies the MSigDB collection abbreviation. E.g.: c("Homo sapiens", "H"). If left empty mouse C7 (immunologic signature) gene set will be used.}

\item{metric_colname}{Name of column which contains the metric used for the ranking of the submitted genelist. "avg_logFC" is used by default.}

\item{pval_adj_cutoff}{Only genes with a more significant adjusted pvalue are considered. Default: 0.001}

\item{Enrichment.Plots}{List of Gene-set names which should be plotted as Enrichment plots in addition to the top 10 Up and Downregulated Genesets.}

\item{my_own_geneset}{A list, where each element contains a gene list and is named with the corresponding pathway name. Default is set to FALSE, so that gene sets from MSigDB are used. Should not contain ".gmt" in name.}
}
\value{
Returns a list containing a tibble with the gene sets and their enrichment scores and Enrichment plots. List element [[1]]: Dataframe with Genesets and statistics. [[2]]: Enrichment plots of top10 Up regulated genesets. [[3]]: Enrichment plots of top10 Down regulated genesets. [[4]]: Enrichment plots of submited gene-sets in parameter Enrichment.Plot.
}
\description{
Conducts a Gene Set Enrichment Analysis (GSEA) on a set of genes submitted in a data frame with a metric each. Works with the output of GEX_genes_cluster or a custom data frame contianing the gene symbols either in a column "symbols" or as rownames and a metric for each gene. The name of the column containing the metric has to be declared via the input metric_colname.
}
\examples{
\dontrun{
gex_gsea <- GEX_GSEA(GEX.cluster.genes.output = df[[1]], MT.Rb.filter = T
, path_to_pathways = "./c5.go.bp.v7.2.symbols.gmt")
cowplot::plot_grid(plotlist=gex_gsea[[2]], ncol=2)
View(gex_gsea[[1]])

#Directly downloading gene set collection from MSigDB to perform gsea
gex_gsea <- GEX_GSEA(GEX.cluster.genes.output =  df[[1]], MT.Rb.filter = T
, path_to_pathways = c("Mus musculus", "C7"))

#Using your own gene list to perform gsea
gex_gsea <- GEX_GSEA(GEX.cluster.genes.output =  df[[1]]
, MT.Rb.filter = T, my_own_geneset = my_geneset)
}
}
