% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Echidna_data.R
\docType{data}
\name{hum_t_h}
\alias{hum_t_h}
\title{hum_t_h}
\format{
A list including 3 elements (data frames): v gene, d gene, j gene,respectively.
\describe{
[[1]]
  \item{gene}{The v gene name}
  \item{seq}{The corresponding sequence}
[[2]]
  \item{gene}{The d gene name}
  \item{seq}{The corresponding sequence}
[[3]]
  \item{gene}{The j gene name}
  \item{seq}{The corresponding sequence}
}
}
\source{
IMGT
}
\usage{
data("hum_t_h")
}
\description{
human germline TRB (heavy chain v,d,j gene segments). When multiple alleles were present, the
first one was included. These names and sequences can be changed by
customized by changing this dataframe. Additionally, repeating elements
can give certain germline gene elements a larger probability of being used
during repertoire evolution.
}
\keyword{datasets}
