% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_anchors.R
\name{initialize_anchors}
\alias{initialize_anchors}
\title{Calculates initial anchor boxes for k-median++ algorithm.}
\usage{
initialize_anchors(annot_df, total_anchors)
}
\arguments{
\item{annot_df}{`data.frame` with widths and heights of bounding boxes.}

\item{total_anchors}{Number of anchors to generate.}
}
\value{
Initial anchor boxes for k-median++ algorithm.
}
\description{
Calculates initial anchor boxes for k-median++ algorithm.
}
