% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_predict.R
\name{intersection_over_union}
\alias{intersection_over_union}
\title{Calculates `Intersection-Over-Union` for two bounding boxes.}
\usage{
intersection_over_union(box1, box2)
}
\arguments{
\item{box1}{Vector `(xmin, ymin, xmax, ymax)` with box coordinates.}

\item{box2}{Vector `(xmin, ymin, xmax, ymax)` with box coordinates.}
}
\value{
`Intersection-Over-Union` for two bounding boxes.
}
\description{
Calculates `Intersection-Over-Union` for two bounding boxes.
}
