% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Echidna_data.R
\docType{data}
\name{iso_SHM_prob}
\alias{iso_SHM_prob}
\title{iso_SHM_prob
A probability dataframe specifying SHM.nuc.prob for cells of different isotypes. The first column is the names of isotypes, while the second column is the SHM.nuc.prob of cell of that isotype. user can define different SHM.nuc.prob for isotypes.}
\format{
a dataframe with 2 columns
}
\usage{
data("iso_SHM_prob")
}
\description{
iso_SHM_prob
A probability dataframe specifying SHM.nuc.prob for cells of different isotypes. The first column is the names of isotypes, while the second column is the SHM.nuc.prob of cell of that isotype. user can define different SHM.nuc.prob for isotypes.
}
\keyword{datasets}
