% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platypus.basicFunctions.R
\name{load.label.data}
\alias{load.label.data}
\title{Load label data, removing the user-specified labels to ignore
TODO This will eventually go away. Also TODO, this should be removing ignore labels... otherwise why bother having it?
TODO Need to remove ignore.label factor level from the data, if it's present :)}
\usage{
load.label.data(fn.labs, classcol.labs, delim = "\\t")
}
\arguments{
\item{fn.labs}{File containing outcome labels}

\item{classcol.labs}{Optional argument. Which column from the labels file to use for learning}

\item{b}{Label class to ignore, if any. Defaults to 'intermediate'}
}
\value{
data frame containing label data
}
\description{
Load label data, removing the user-specified labels to ignore
TODO This will eventually go away. Also TODO, this should be removing ignore labels... otherwise why bother having it?
TODO Need to remove ignore.label factor level from the data, if it's present :)
}
