% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platypus.basicFunctions.R
\name{load.parameterfile}
\alias{load.parameterfile}
\title{Create a view from a configuration file}
\usage{
load.parameterfile(filename, delim = "\\t")
}
\arguments{
\item{filename}{File containing config information}

\item{delim}{Optional delimeter for the file containing view data (not the config file!)}
}
\value{
View object
}
\description{
Create a view from a configuration file
}
\keyword{platypus}
