% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_darknet_weights.R
\name{load_darknet_weights}
\alias{load_darknet_weights}
\title{Loads `Yolo3 Darknet` (https://pjreddie.com/darknet/yolo/) weights into \code{\link[platypus]{yolo3}} model.}
\usage{
load_darknet_weights(model, weights_path)
}
\arguments{
\item{model}{\code{\link[platypus]{yolo3}} model.}

\item{weights_path}{Filepath to weights file (You can download valid weights file from "https://pjreddie.com/media/files/yolov3.weights").}
}
\value{
\code{\link[platypus]{yolo3}} model with pretrained weights.
}
\description{
Loads `Yolo3 Darknet` (https://pjreddie.com/darknet/yolo/) weights into \code{\link[platypus]{yolo3}} model.
}
