% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Echidna_data.R
\docType{data}
\name{mus_b_trans}
\alias{mus_b_trans}
\title{mus_b_trans
A data frame contains mouse B cell average gene expression for multiple cell types, with the rows representing the gene names, column names representing the cell type names.
The original single cell sequencing data is retrieved from 10xgenomics and combined with experimental data from.#?
The expression level for different cell types are obtained by calculating the average expression after sorting the original data by markers as shown below.
\describe{
\item{NaiveBcell}{Cd19+;Cd27-;Cd38-}
\item{GerminalcenterBcell}{Fas+;Cd19+}
\item{Plasmacell}{Sdc1+}
\item{MemoryBcell}{Cd38+;Fas-}
}}
\format{
A data frame with 26538 rows and 4 variables, with the rows representing the gene names, column names representing the cell type names.
}
\source{
https://support.10xgenomics.com/single-cell-vdj/datasets/3.0.0/vdj_v1_mm_c57bl6_pbmc_5gex
https://support.10xgenomics.com/single-cell-vdj/datasets/3.0.0/vdj_v1_mm_balbc_pbmc_5gex
}
\usage{
data("mus_b_trans")
}
\description{
mus_b_trans
A data frame contains mouse B cell average gene expression for multiple cell types, with the rows representing the gene names, column names representing the cell type names.
The original single cell sequencing data is retrieved from 10xgenomics and combined with experimental data from.#?
The expression level for different cell types are obtained by calculating the average expression after sorting the original data by markers as shown below.
\describe{
\item{NaiveBcell}{Cd19+;Cd27-;Cd38-}
\item{GerminalcenterBcell}{Fas+;Cd19+}
\item{Plasmacell}{Sdc1+}
\item{MemoryBcell}{Cd38+;Fas-}
}
}
\keyword{datasets}
