% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbForests_data.R
\docType{data}
\name{new}
\alias{new}
\title{Example repertoire for testing and examples}
\format{
An object of class \code{list} of length 1662.
}
\usage{
new
}
\description{
Example repertoire for testing and examples
}
\keyword{datasets}
