% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_predict.R
\name{non_max_suppression}
\alias{non_max_suppression}
\title{Applies `Non-Maximum-Suppression` for a list of bounding boxes.}
\usage{
non_max_suppression(boxes, n_class, nms_threshold)
}
\arguments{
\item{boxes}{List of bounding boxes. Each box is a vector in format `(xmin, ymin, xmax, ymax)`.}

\item{n_class}{Number of prediction classes.}

\item{nms_threshold}{`Non-Maximum-Suppression` threshold.}
}
\value{
List of non-overlapping bounding boxes.
}
\description{
Applies `Non-Maximum-Suppression` for a list of bounding boxes.
}
