% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Echidna_data.R
\docType{data}
\name{pheno_SHM_prob}
\alias{pheno_SHM_prob}
\title{pheno_SHM_prob
A probability dataframe specifying SHM.nuc.prob for cells of different phenotypes. The first column is the names of phenotypes, while the second column is the SHM.nuc.prob of cell of that phenotype. user can define different SHM.nuc.prob for phenotypes.}
\format{
a dataframe with 2 columns
}
\usage{
data("pheno_SHM_prob")
}
\description{
pheno_SHM_prob
A probability dataframe specifying SHM.nuc.prob for cells of different phenotypes. The first column is the names of phenotypes, while the second column is the SHM.nuc.prob of cell of that phenotype. user can define different SHM.nuc.prob for phenotypes.
}
\keyword{datasets}
