% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platypus.basicFunctions.R
\name{platypus.predict}
\alias{platypus.predict}
\title{Given a trained platypus model, make predictions. Generally used internally by platypus}
\usage{
platypus.predict(view.list, majority, test.ids, unique.labels, labels,
  join.fxn = "ensemble")
}
\arguments{
\item{view.list}{List of views to use for predictions}

\item{majority}{Confidence threshold for predictions}

\item{test.ids}{Samples to be predicted}

\item{unique.labels}{Class labels}

\item{labels}{Labels for known samples}

\item{join.fxn}{Function used to make group votes}
}
\value{
Matrix of predictions (One per view, overall, etc)
}
\description{
Given a trained platypus model, make predictions. Generally used internally by platypus
}
