% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlatypusML_balance.R
\name{PlatypusML_balance}
\alias{PlatypusML_balance}
\title{This PlatypusML_classification function takes as input encoded features obtained using the PlatypusML_extract_features function. The function runs cross validation on a specified number of folds 
for different classification models and reports the AUC scores and ROC curves.}
\usage{
PlatypusML_balance(matrix, label.1, label.2, proportion, random.seed)
}
\arguments{
\item{matrix}{Matrix. Output of the PlatypusML_extract_features function, with the last column storing the label.}

\item{label.1}{String. The label of the first class.}

\item{label.2}{String. The label of the second class.}

\item{random.seed}{Integer. The seed to be set when sampling for balancing the dataset.}

\item{proportion.}{Vector of size 2 (floats between 0 and 1 that need to sum up to 1). Specifies the proportions for the two classes.
The smaller proportion will be assigned to the minority class by default.}
}
\value{
This function returns a matrix containing equal number of samples for the two classes.
#' @examples
\dontrun{
TODO: example
}
}
\description{
This PlatypusML_classification function takes as input encoded features obtained using the PlatypusML_extract_features function. The function runs cross validation on a specified number of folds 
for different classification models and reports the AUC scores and ROC curves.
}
