% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cv.R
\name{plot.cv}
\alias{plot.cv}
\title{Given expanded output from CV tests, generate plots showing AUC change & labels learned.}
\usage{
plot.cv(fn.labs, folder, plot.single.cvfolds = FALSE)
}
\arguments{
\item{fn.labs}{File containing outcome labels}

\item{folder}{Output directory where results will be written and cv.platypus results already are}

\item{plot.single.cvfolds}{Whether or not to plot lines for each cv fold, in addition to the general line plot}
}
\value{
Nothing. Plots are printed to file in user-provided folder
}
\description{
Given expanded output from CV tests, generate plots showing AUC change & labels learned.
}
