% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.llv.R
\name{plot.llv}
\alias{plot.llv}
\title{Given expanded output from CV and LLV tests, generate a label learning matrix and plot results}
\usage{
\method{plot}{llv}(fn.labs, folder)
}
\arguments{
\item{fn.labs}{File containing outcome labels}

\item{folder}{Output directory where results will be written and cv.platypus results already are}
}
\value{
Nothing. Plots are printed to file in user-provided folder
}
\description{
Given expanded output from CV and LLV tests, generate a label learning matrix and plot results
}
