% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_raster}
\alias{plot_raster}
\title{Generates raster image.}
\usage{
plot_raster(plot_data, grayscale)
}
\arguments{
\item{plot_data}{`data.frame` with `x`, `y` coordinates and color values.}

\item{grayscale}{Should images be plotted in grayscale.}
}
\value{
Raster image.
}
\description{
Generates raster image.
}
