% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_rgb_raster}
\alias{plot_rgb_raster}
\title{Generates raster image.}
\usage{
plot_rgb_raster(plot_data)
}
\arguments{
\item{plot_data}{`data.frame` with `x`, `y` coordinates and color values.}
}
\value{
Raster image.
}
\description{
Generates raster image.
}
