% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_cell_analysis.R
\name{public_cell_analysis}
\alias{public_cell_analysis}
\title{public_cell_analysis function:
Creates Plots using the find_public_cells() output vgm. Which type of plot and analysis should be done can be chosen by selecting the desired analysis.type input. Default is set to "overlap type counts"}
\usage{
public_cell_analysis(
  normalize_y_n,
  find_public_cells_output,
  analysis.type,
  include.light.chain,
  save.plot,
  nr_clones
)
}
\arguments{
\item{find_public_cells_output}{Object, output of find_public_cells function}

\item{analysis.type}{Character, defines what analysis should be performed. Options are: "overlap type counts", "public cell numbers", "public clone numbers",  "clonal expansion by overlap type", "isotype usage of public cells" and "transcriptional cluster usage of public cells". Default is "overlap type counts".}

\item{inlcude.light.chain}{Logical, default is FALSE. Only needed if analysis.type = "clonal expansion by overlap type". If set to TRUE, also cdr3_aa light chain overlaps are colored in the created clonal expansion plots.}

\item{save.pot}{Logical, default is FALSE. If set to TRUE, created plots are saved in the current working directory.}
}
\description{
1. Plot: Overlap Type Counts per Sample (CDRH3, CDRL3 and CDRL3.CDRH3 overlaps) - analysis.type= "overlap type counts"
}
\details{
2. Plot: Public Cell Numbers - analysis.type = "public cell numbers"
Note: It is likely that a cell which has a CDRH3 overlap has a CDRL3 overlap with a cell in the public data as well, but is not listed as a CDRL3_CDRH3_aa overlap. Because cells are only assigned a complete overlap if both (heavy and light chain) overlaps refer to the same cell in the public data.
Some CDRL3 overlaps might therefore be overwritten with CDRH3 overlaps due to their higher significance.

3. Plot: Public Clone Numbers - analysis.type = "public clone numbers"

4. Plot: Clonal expansion colored by overlap type - analysis.type = "clonal expansion by overlap type"
Note: Here you can use an additional input called: include.light.chain
If include.light.chain = TRUE, also light chain overlaps are colored in the clonal expansion plots. Default is set to FALSE
Also, this function returns a pdf of clonal expansion plots (all samples) instead of only one or two plots, if save.plot = TRUE

5. Isotype usage of public cells - analysis.type = "isotype usage of public cells"
Creates stacked barplots of the isotype usage of all public cells

6. Transcriptional cluster usage of public cells - analysis.type = "transcriptional cluster usage of public cells"
Creates stacked barplots of the transcriptional cluster usage of public cells
}
