% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_generator.R
\name{read_annotations_from_labelme}
\alias{read_annotations_from_labelme}
\title{Reads bounding box annotations from JSON files. Each XML has to be in `labelme JSON` format.}
\usage{
read_annotations_from_labelme(annot_paths, indices, images_path, labels)
}
\arguments{
\item{annot_paths}{List to XML annotations filepaths.}

\item{indices}{Indices specifying which files to read. If `NULL` all files are loaded.}

\item{images_path}{Path to directory with images.}

\item{labels}{Character vector containing class labels. For example \code{\link[platypus]{coco_labels}}.}
}
\value{
List of bounding box coordinates, heights, widths and image filepaths.
}
\description{
Reads bounding box annotations from JSON files. Each JSON has to be in `labelme JSON` format.
}
