% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_generator.R
\name{read_images_from_directory}
\alias{read_images_from_directory}
\title{Reads images from directory.}
\usage{
read_images_from_directory(paths, indices = NULL, target_size = c(256,
  256), grayscale = FALSE, scale = 1/255)
}
\arguments{
\item{paths}{Images directories.}

\item{indices}{Indices of selected images. If `NULL` all images in `paths` will be selected.}

\item{target_size}{Images size (height, width). Default to `c(256, 256)`.}

\item{grayscale}{Boolean, whether to load the image as grayscale.}

\item{scale}{Scaling factor for images pixel values. Default to `1 / 255`.}
}
\description{
Reads images from directory.
}
