% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3.R
\name{reshape_yolo3_output}
\alias{reshape_yolo3_output}
\title{Reshapes Yolo3 output grid.}
\usage{
reshape_yolo3_output(x, anchors_per_grid, n_class)
}
\arguments{
\item{x}{Yolo3 grid output layer object.}

\item{anchors_per_grid}{Number of anchors/boxes per one output grid.}

\item{n_class}{Number of prediction classes.}
}
\value{
Reshaped Yolo3 output grid.
}
\description{
Reshapes Yolo3 output grid from `(S, H, W, anchors_per_grid * (5 + n_class))` to `(S, H, W, anchors_per_grid, 5 + n_class)`.
}
