% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_fit.R
\name{segmentation_fit_generator}
\alias{segmentation_fit_generator}
\title{Fits `segmentation` model using data generator.}
\usage{
segmentation_fit_generator(
  model,
  generator,
  epochs,
  steps_per_epoch,
  validation_generator = NULL,
  validation_steps_per_epoch = NULL,
  model_filepath = NULL
)
}
\arguments{
\item{model}{`segmentation` Model.}

\item{generator}{`segmentation` Data generator.}

\item{epochs}{Number of epochs.}

\item{steps_per_epoch}{Steps per epoch.}

\item{validation_generator}{Validation data generator.}

\item{validation_steps_per_epoch}{Validation steps per epoch.}

\item{model_filepath}{Path to save the model.}
}
\description{
Fits `segmentation` model using data generator.
}
