% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_generator.R
\name{segmentation_generator}
\alias{segmentation_generator}
\title{Generates batches of data (images and masks). The data will be looped over (in batches).}
\usage{
segmentation_generator(
  path,
  colormap,
  mode = "dir",
  only_images = FALSE,
  net_h = 256,
  net_w = 256,
  grayscale = FALSE,
  scale = 1/255,
  batch_size = 32,
  shuffle = TRUE,
  subdirs = c("images", "masks"),
  column_sep = ";"
)
}
\arguments{
\item{path}{Images and masks directory.}

\item{colormap}{Class color map. For example \code{\link[platypus]{voc_colormap}}.}

\item{mode}{Character. One of `c("dir", "nested_dirs", "config_file")`}

\item{only_images}{Should generator read only images (e.g. on train set for predictions).}

\item{net_h}{Input layer height. Must be equal to `2^x, x - natural`.}

\item{net_w}{Input layer width. Must be equal to `2^x, x - natural`.}

\item{grayscale}{Defines input layer color channels -  `1` if `TRUE`, `3` if `FALSE`.}

\item{scale}{Scaling factor for images pixel values. Default to `1 / 255`.}

\item{batch_size}{Batch size.}

\item{shuffle}{Should data be shuffled.}

\item{subdirs}{Vector of two characters containing names of subdirectories with images and masks.}

\item{column_sep}{Character. Configuration file separator.}
}
\description{
Generates batches of data (images and masks). The data will be looped over (in batches).
}
