% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_fit.R
\name{segmentation_predict_generator}
\alias{segmentation_predict_generator}
\title{Calculates predictions on new samples using `segmentation` data generator.}
\usage{
segmentation_predict_generator(model, generator, steps)
}
\arguments{
\item{model}{`segmentation` Model.}

\item{generator}{`segmentation` Data generator.}

\item{steps}{Steps in epoch.}
}
\description{
Calculates predictions on new samples using `segmentation` data generator.
}
