% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.predictor.R
\name{single.elasticNet.predictor}
\alias{single.elasticNet.predictor}
\title{Grid search of parameters for an elastic net model using data X and labels y.}
\usage{
single.elasticNet.predictor(X, y, alpha = seq(0, 1, 0.1),
  iterations = 10, nfolds = 10, measure = "auc", viewName = "",
  ignore.label = "intermediate")
}
\arguments{
\item{X}{Feature matrix (sample per row)}

\item{y}{Named vector of labels}

\item{alpha}{Values of alpha to test}

\item{iterations}{Number of iterations per alpha to run the test}

\item{nfolds}{Number of folds to use}

\item{measure}{Model performance measure to use. Default AUC.}

\item{viewName}{Currently not in use, will be name of the view created from this.}

\item{ignore.label}{Label type to be excluded (samples with this label will not be used in training).}
}
\value{
list containing alpha and accuracy (or whichever measure selected) for best model.
}
\description{
Grid search of parameters for an elastic net model using data X and labels y.
}
\examples{
X <- matrix(rnorm(1000), nrow=100)
y <- c(rep('MOO',50),rep('OINK',50))
names(y) <- paste0('Sample',seq(nrow(X)))
rownames(X) <- paste0('Sample',seq(nrow(X)))
res <- single.elasticNet.predictor(X,y)

}
