% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.predictor.R
\name{single.predictor}
\alias{single.predictor}
\title{Parameter sweeps of svm, random forest, and elastic net models for config generation.}
\usage{
single.predictor(X, y, model, ignore.label = "intermediate")
}
\arguments{
\item{X}{Feature matrix (sample per row)}

\item{y}{Named vector of labels}

\item{model}{Type of single predictor model to build. Choose from en (elastic net), rf (random forest), or svm (support vector machine)}

\item{ignore.label}{Label class to ignore. Will be removed from labels set before training.}
}
\value{
list containing optimal model summary
}
\description{
Parameter sweeps of svm, random forest, and elastic net models for config generation.
}
\examples{
X <- matrix(rnorm(10000), nrow=100)
y <- c(rep('MOO',50),rep('OINK',50))
names(y) <- paste0('Sample',seq(nrow(X)))
rownames(X) <- paste0('Sample',seq(nrow(X)))
colnames(X) <- paste0('Feature',seq(ncol(X)))
y <- factor(y)
res <- single.predictor(X,y,model='rf')
res <- single.predictor(X,y,model='en')
res <- single.predictor(X,y,model='svm')

}
