% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_predictor.R
\name{single.randomForest.predictor}
\alias{single.randomForest.predictor}
\title{Grid search of parameters for a random forest model using data X and labels y.}
\usage{
single.randomForest.predictor(X, y, mtry = NA, ntree = c(500, 1000,
  1500, 2000))
}
\arguments{
\item{X}{Feature matrix (sample per row)}

\item{y}{Named vector of labels}

\item{mtry}{Number of predictors samples for splitting at each node}

\item{ntree}{Number of trees grown}

\item{iterations}{Number of iterations per alpha to run the test}

\item{nfolds}{Number of folds to use}

\item{measure}{Model performance measure to use. Defaults to accuracy.}

\item{ignore.label}{Label type to be excluded (samples with this label will not be used in training).}
}
\value{
list containing mtry, ntree, and accuracy (or whichever measure selected) for best model.
}
\description{
Grid search of parameters for a random forest model using data X and labels y.
}
\examples{
X <- matrix(rnorm(1000), nrow=100)
y <- c(rep('MOO',50),rep('OINK',50))
names(y) <- paste0('Sample',seq(nrow(X)))
rownames(X) <- paste0('Sample',seq(nrow(X)))
res <- single.randomForest.predictor(X,y)

}
