% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Platypus_data.R
\docType{data}
\name{small_vgm}
\alias{small_vgm}
\title{Small VDJ GEX matrix (VGM) for function testing purposes}
\format{
An object of class \code{list} of length 5.
}
\usage{
small_vgm
}
\description{
Small VDJ GEX matrix (VGM) for function testing purposes
}
\references{
R package Platypus : https://doi.org/10.1093/nargab/lqab023
}
\keyword{datasets}
