% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Echidna_data.R
\docType{data}
\name{special_v}
\alias{special_v}
\title{special_v
a dataframe, of heavy and light chain v gene combination and their probability to be selected for expansion.}
\format{
An object of class \code{data.frame} with 5 rows and 3 columns.
}
\usage{
data("special_v")
}
\description{
special_v
a dataframe, of heavy and light chain v gene combination and their probability to be selected for expansion.
}
\keyword{datasets}
