% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_generator.R
\name{split_masks_into_binary}
\alias{split_masks_into_binary}
\title{Splits multi-class mask into binary masks.}
\usage{
split_masks_into_binary(mask, colormap)
}
\arguments{
\item{mask}{Segmentation mask.}

\item{colormap}{Class color map. For example \code{\link[platypus]{voc_colormap}}.}
}
\description{
Splits multi-class mask into binary masks.
}
