% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Echidna_data.R
\docType{data}
\name{trans_switch_prob_t}
\alias{trans_switch_prob_t}
\title{trans_switch_prob_t
The probability for T cell transcriptome states switching. The row names of the matrix are the cell states the cell is switching from, the column names are the cells states the cell is switching to.}
\format{
A 7*7 matrix. The row and clumn names are: "NaiveCd4","ActivatedCd4","MemoryCd4","NaiveCd8","EffectorCd8","MemoryCd8","ExhaustedCd8".
}
\usage{
data("trans_switch_prob_t")
}
\description{
trans_switch_prob_t
The probability for T cell transcriptome states switching. The row names of the matrix are the cell states the cell is switching from, the column names are the cells states the cell is switching to.
}
\keyword{datasets}
