% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_loss.R
\name{transform_box_to_min_max}
\alias{transform_box_to_min_max}
\title{Transforms boxes to min/max coordinates.}
\usage{
transform_box_to_min_max(box)
}
\arguments{
\item{box}{Boxes coordinates.}
}
\value{
Boxes min/max coordinates.
}
\description{
Transforms boxes to min/max coordinates.
}
