% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_net.R
\name{u_net_double_conv2d}
\alias{u_net_double_conv2d}
\title{Creates a double convolutional U-Net block.}
\usage{
u_net_double_conv2d(
  input,
  filters,
  kernel_size,
  batch_normalization = TRUE,
  kernel_initializer = "he_normal"
)
}
\arguments{
\item{input}{Model or layer object.}

\item{filters}{Integer, the dimensionality of the output space (i.e. the number of output filters in the convolution).}

\item{kernel_size}{An integer or list of 2 integers, specifying the width and height of the 2D convolution window. Can be a single integer to specify the same value for all spatial dimensions.}

\item{batch_normalization}{Should batch normalization be used in the block.}

\item{kernel_initializer}{Initializer for the kernel weights matrix.}
}
\value{
Double convolutional U-Net block.
}
\description{
Creates a double convolutional U-Net block.
}
