% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Echidna_other_functions.R
\name{umap.top.highlight}
\alias{umap.top.highlight}
\title{Set idents for top abundant clones in Seurat object, get ready for highlight the top abundant clones in UMAP.}
\usage{
umap.top.highlight(gex, all.contig.annotations, top.n)
}
\arguments{
\item{gex}{output from get.umap function.}

\item{all.contig.annotations}{The output dataframe all_contig_annotations from simulation.}

\item{top.n}{The top n abundant clones to be shown in the plot. If missing, all clones will be shown.}
}
\value{
a Seurat object ready for highlight the top abundant clones in UMAP
}
\description{
Set idents for top abundant clones in Seurat object, get ready for highlight the top abundant clones in UMAP.
}
