% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_predict.R
\name{unite_binary_masks}
\alias{unite_binary_masks}
\title{Unites binary masks into multi-class mask.}
\usage{
unite_binary_masks(masks, colormap)
}
\arguments{
\item{masks}{Segmentation masks.}

\item{colormap}{Class color map. For example \code{\link[platypus]{voc_colormap}}.}
}
\description{
Unites binary masks into multi-class mask.
}
