% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ_clonotype_phylo.R
\name{VDJ_clonotype_phylo}
\alias{VDJ_clonotype_phylo}
\title{Returns a list of clonotype dataframes following additional clonotyping. This function works best following filtering to ensure that each clone only has one heavy chain and one light chain.}
\usage{
VDJ_clonotype_phylo(clonotype.list, clone.strategy, homology.threshold)
}
\arguments{
\item{clonotype.list}{Output from VDJ_analyze function. This should be a list of clonotype dataframes, with each list element corresponding to a single VDJ repertoire.}

\item{clone.strategy}{String describing the clonotyping strategy. Possible options include 'cdr3.aa','hvj.lvj','hvj.lvj.cdr3lengths','Hvj.Lvj.CDR3length.CDR3homology', 'Hvj.Lvj.CDR3length.CDRH3homology', 'CDR3homology',or 'CDRH3homology'. 'CDR3aa' will convert the default cell ranger clonotyping to amino acid based. 'Hvj.Lvj' groups B cells with identical germline genes (V and J segments for both heavy chain and light chain. Those arguments including 'CDR3length' will group all sequences with identical CDRH3 and CDRL3 sequence lengths. Those arguments including 'CDR3homology' will additionally impose a homology requirement for CDRH3 and CDRL3 sequences.'CDR3homology',or 'CDRH3homology' will group sequences based on homology only (either of the whole CDR3 sequence or of the CDRH3 sequence respictevely).
All homology calculations are performed on the amino acid level.}

\item{homology.threshold}{Numeric value between 0 and 1 corresponding to the homology threshold forn the clone.strategy arguments that require a homology threshold. Default value is set to 70 percent sequence homology. For 70 percent homology, 0.3 should be supplied as input.}
}
\value{
list_dfs Returns a list of clonotype dataframes. list_dfs[[1]] is the first element of the list, etc.
}
\description{
Returns a list of clonotype dataframes following additional clonotyping. This function works best following filtering to ensure that each clone only has one heavy chain and one light chain.
}
\examples{
\dontrun{
VDJ_clonotype(clonotype.list=VDJ.analyze.output,
 clone.strategy="Hvj.Lvj.CDR3length.CDR3homology",
 homology.threshold=".3")
 }
}
