% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ_logoplot.R
\name{VDJ_logoplot}
\alias{VDJ_logoplot}
\title{Plots a logoplot of the CDR3 aminoacid region ! For platypus v3 and more flexibility use VDJ_logoplot_vector}
\usage{
VDJ_logoplot(VDJ.object, length_cdr3)
}
\arguments{
\item{VDJ.object}{Output of the VDJ_analyze function}

\item{length_cdr3}{Integer indicating the length of the CDR3 regions that are selected to be plotted.}
}
\value{
Returns the logo plot.
}
\description{
Plots a logoplot of the CDR3 aminoacid region ! For platypus v3 and more flexibility use VDJ_logoplot_vector
}
\examples{
\dontrun{
VDJ_logoplot <- function(VDJ.object = VDJ_analyze.out,length_cdr3 = 10)
}
}
