% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ_overlap_heatmap.R
\name{VDJ_overlap_heatmap}
\alias{VDJ_overlap_heatmap}
\title{Yields overlap heatmap and datatable of features or combined features for different samples or groups}
\usage{
VDJ_overlap_heatmap(
  VDJ,
  feature.columns,
  grouping.column,
  plot.type,
  pvalues.label.size,
  axis.label.size,
  add.barcode.table,
  platypus.version
)
}
\arguments{
\item{VDJ}{VDJ output of the VDJ_GEX_matrix function (VDJ_GEX_matrix.output[[1]])}

\item{feature.columns}{A character array of column names of which the overlap should be displayed. The content of these columns is pasted together (separated by "/"). E.g. if the overlap in cells germline gene usage is desired, the input could be c("VDJ_jgene","VDJ_dgene","VDJ_vgene"). These columns would be pasted and compared across the grouping variable.}

\item{grouping.column}{A column which acts as a grouping variable. If repertoires are to be compared use the sample_id column.}

\item{plot.type}{Character. Either "ggplot" or "pheatmap". Defaults to Pheatmap}

\item{pvalues.label.size}{Numeric. Defaults to 4. Is passed on to ggplot theme}

\item{axis.label.size}{Numeric. Defaults to 4. Is passed on to ggplot theme}

\item{add.barcode.table}{Boolean. Defaults to T. Whether to generate a dataframe with frequencies and barcodes of cells with overlapping features. This is useful to e.g. analyze deferentially expressed genes between cells of two samples or groups expressing the same VDJ or VJ chain}

\item{platypus.version}{Character. At the moment this function runs only on the output of the VDJ_GEX_matrix function meaning that it is exclusively part of Platypus "v3". With further updates the functionality will be extended.}
}
\value{
A list of a ggplot (out[[1]]), the source table or matrix for the plot out[[2]] and a table containing additional information in case that add.barcode.table was set to TRUE (out[[3]])
}
\description{
Yields overlap heatmap and datatable of features or combined features for different samples or groups
}
\examples{
\dontrun{
#To test the overlap of barcodes between multiple samples
barcode_overlap <- VDJ_overlap_heatmap(VDJ = VDJ_comb[[1]]
,feature.columns = c("barcode"),
grouping.column = "sample_id", axis.label.size = 15)
}
}
