% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ_per_cell_matrix.R
\name{VDJ_per_cell_matrix}
\alias{VDJ_per_cell_matrix}
\title{Processes and organizes the repertoire sequening data from cellranger vdj and returns a single matrix where the rows are individual cells and the columns are repertoire features.}
\usage{
VDJ_per_cell_matrix(VDJ.out.directory)
}
\arguments{
\item{VDJ.out.directory}{Character vector with each element containing the path to the output of cellranger vdj runs. Multiple repertoires to be integrated in a single transcriptome should be supplied as multiple elements of the character vector. This can be left blank if supplying the clonotypes and contig files directly as input. This pipeline assumes that the output file names have not been changed from the default 10x settings in the /outs/ folder. This is compatible with B and T cell repertoires (both separately and simultaneously).}
}
\value{
returns a single matrix where the rows are individual cells and the columns are repertoire features.
}
\description{
Processes and organizes the repertoire sequening data from cellranger vdj and returns a single matrix where the rows are individual cells and the columns are repertoire features.
}
\examples{
\dontrun{
example.vdj.analyze <- VDJ_per_cell_matrix(VDJ.out.directory = "~/path/to/cellranger/vdj/outs/")
}

}
