% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ_plot_SHM.R
\name{VDJ_plot_SHM}
\alias{VDJ_plot_SHM}
\title{Plots for SHM based on MIXCR output generated using the VDJ_call_MIXCR function and appended to the VDJ.GEX.matrix.output}
\usage{
VDJ_plot_SHM(
  VDJ.mixcr.matrix,
  group.by,
  quantile.label,
  point.size,
  mean.line.color,
  platypus.version
)
}
\arguments{
\item{VDJ.mixcr.matrix}{Output dataframe from the VDJ_call_MIXCR function or a dataframe generated using the VDJ_GEX_matrix function and supplemented with MIXCR information}

\item{group.by}{Character. Defaults to "sample_id". Column name of VDJ.matrix to split VDJ.matrix by. For each unique entry in that column a set of plots will be generated. This can be useful to plot SHM by expansion or by transcriptomics-derived clusters}

\item{quantile.label}{Numeric. Defaults to 0.9. Which points to label in the SHM scatterplot. If set to 0.9, the top 10\% of cells by SHM number will be labelled. If ggrepel throws a warning, concerning overlap it is recommended to attempt to lable less points to avoid cluttering}

\item{point.size}{Size of points in plots. Passed to geom_jitter()}

\item{mean.line.color}{Color of mean bar in dotplots. Passed to geom_errorbar()}

\item{platypus.version}{Character. Only "v3" available.}
}
\value{
Returns a list of ggplot objects. out\[\[1\]\] is a boxplot comparing SHM by group.by. out\[\[2\]\] to out\[\[n\]\] are plots for each group that visualize VDJ and VJ SHM distribution for each group. Data for any plot can be accessed via out \[\[any\]\]$data
}
\description{
Plots for SHM based on MIXCR output generated using the VDJ_call_MIXCR function and appended to the VDJ.GEX.matrix.output
}
\examples{
\dontrun{

Standard plots
SHM_plots <- VDJ_plot_SHM(VDJ.matrix = VDJ_call_MIXCR.output, group.by = "sample_id", quantile.label = 0.9)

Group by transcriptional cluster and label only top 1\\%
SHM_plots <- VDJ_plot_SHM(VDJ.matrix = VDJ_call_MIXCR.output, group.by = "seurat_clusters", quantile.label = 0.99)
}
}
