% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platypus.basicFunctions.R
\name{view.train.ElasticNet}
\alias{view.train.ElasticNet}
\title{Train an Elastic Net View
TODO This is an internal fxn for platypus right, so probably shouldn't share it with users?}
\usage{
\method{view.train}{ElasticNet}(labels, view.object, nfolds = 10)
}
\arguments{
\item{labels}{label data}

\item{view.object}{A view object}

\item{nfolds}{Number of cross-validation folds}
}
\value{
Newly re-trained view object
}
\description{
Train an Elastic Net View
TODO This is an internal fxn for platypus right, so probably shouldn't share it with users?
}
