% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platypus.basicFunctions.R
\name{view.train.RandomForest}
\alias{view.train.RandomForest}
\title{Train a Random Forest View
TODO This is an internal fxn for platypus right, so probably shouldn't share it with users?}
\usage{
\method{view.train}{RandomForest}(labels, view.object)
}
\arguments{
\item{labels}{label data}

\item{view.object}{A view object}
}
\value{
Newly re-trained view object
}
\description{
Train a Random Forest View
TODO This is an internal fxn for platypus right, so probably shouldn't share it with users?
}
