% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ_alpha_beta_Vgene_circos.R
\name{VJ_alpha_beta_Vgene_circos}
\alias{VJ_alpha_beta_Vgene_circos}
\title{Makes a Circos plot from the VDJ_analyze output. Connects the V-alpha with the corresponding V-beta gene for each clonotype.}
\usage{
VJ_alpha_beta_Vgene_circos(
  VDJ.analyze.output,
  V.or.J,
  B.or.Tcells,
  label.threshold,
  c.threshold,
  cell.level,
  clonotype.per.gene.threshold
)
}
\arguments{
\item{VDJ.analyze.output}{The output of the VDJ_analyze function. A list of data frames for each sample containing the clonotype informations.}

\item{V.or.J}{Determines whether to plot the alpha beta gene pairing of the V or J genes. "V", "J" or "both" as possible inputs. Default: "both".}

\item{B.or.Tcells}{Specify whether B or T cells are being analyzed ("B" or "T"). If not specified, function attempts to decide based on gene names.}

\item{label.threshold}{Minimal amount of clonotypes per gene neccessary to add a gene label to the sector. Default: 0.}

\item{c.threshold}{Only clonotypes are considered with a frequency higher then c.threshold. Allows to filter for only highly expanded clonotypes.}

\item{cell.level}{Logical, defines whether weight of connection should be based on number of clonotypes of number of cells. Default: number of clonotypes.}

\item{clonotype.per.gene.threshold}{How many clonotypes are required to plot a sector for a gene. Filters the rows and colums of the final adjacency matrix.}
}
\value{
Returns list of plots. The first n elements contain the circos plot of the n datasets from the VDJ.analyze function. The n+1 element contains a list of the n adjancey matrices for each dataset.
}
\description{
Makes a Circos plot from the VDJ_analyze output. Connects the V-alpha with the corresponding V-beta gene for each clonotype.
}
\examples{
\dontrun{
 plots <- VJ_alpha_beta_Vgene_circos(vdj_repertoire_tcells)
}
}
