% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_net.R
\docType{data}
\name{voc_colormap}
\alias{voc_colormap}
\title{VOC dataset segmentation color map.}
\format{
An object of class \code{list} of length 21.
}
\usage{
voc_colormap
}
\value{
VOC dataset segmentation color map.
}
\description{
VOC dataset segmentation color map.
}
\keyword{datasets}
