% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_net.R
\docType{data}
\name{voc_labels}
\alias{voc_labels}
\title{VOC dataset labels.}
\format{
An object of class \code{character} of length 21.
}
\usage{
voc_labels
}
\value{
VOC dataset labels.
}
\description{
VOC dataset labels.
}
\keyword{datasets}
