% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.config.R
\name{write.config}
\alias{write.config}
\title{Write a configuration to file. Generally used by gen.config()}
\usage{
write.config(x, v, task, fn.config = "config_TEST.txt", view.name = NA)
}
\arguments{
\item{x}{Parameter sweep results generated by single.predictor()}

\item{v}{View data frame}

\item{task}{Task name}

\item{fn.config}{Filename of the config to be created}
}
\description{
Write a configuration to file. Generally used by gen.config()
}
