% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3.R
\name{yolo3_conv2d}
\alias{yolo3_conv2d}
\title{Creates a convolutional Yolo3 unit.}
\usage{
yolo3_conv2d(inputs, filters, name)
}
\arguments{
\item{filters}{Integer, the dimensionality of the output space (i.e. the number of output filters in the convolution).}

\item{name}{Model name.}

\item{input}{Model or layer object.}
}
\value{
Convolutional Yolo3 unit.
}
\description{
Creates a convolutional Yolo3 unit.
}
