% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_train_on_batch.R
\name{yolo3_fit_generator}
\alias{yolo3_fit_generator}
\title{Fits `YOLOv3` model using data generator.}
\usage{
yolo3_fit_generator(
  model,
  generator,
  epochs,
  steps_per_epoch,
  validation_generator = NULL,
  validation_steps_per_epoch = NULL,
  model_filepath = NULL,
  save_best_only = TRUE,
  monitor = "loss"
)
}
\arguments{
\item{model}{`YOLOv3` Model.}

\item{generator}{`YOLOv3` Data generator.}

\item{epochs}{Number of epochs.}

\item{steps_per_epoch}{Steps per epoch.}

\item{validation_generator}{Validation data generator.}

\item{validation_steps_per_epoch}{Validation steps per epoch.}

\item{model_filepath}{Path to save the model.}

\item{save_best_only}{If set to `TRUE` model will be saved only if selected `metric` was better than in previous epochs.}

\item{monitor}{Metric selected to be monitored.}
}
\description{
Fits `YOLOv3` model using data generator.
}
