% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_generator.R
\name{yolo3_generator}
\alias{yolo3_generator}
\title{Generates batches of data (images and box coordinates/scores). The data will be looped over (in batches).}
\usage{
yolo3_generator(
  annot_path,
  images_path,
  only_images = FALSE,
  net_h = 416,
  net_w = 416,
  annot_format = "pascal_voc",
  grayscale = FALSE,
  scale = 1/255,
  anchors = coco_anchors,
  labels = coco_labels,
  batch_size = 32,
  shuffle = TRUE
)
}
\arguments{
\item{annot_path}{Annotations directory.}

\item{images_path}{Images directory.}

\item{only_images}{Should generator read only images (e.g. on train set for predictions).}

\item{net_h}{Input layer height. Must be divisible by `32`.}

\item{net_w}{Input layer width. Must be divisible by `32`.}

\item{annot_format}{Annotations format. One of `pascal_voc`, `labelme`.}

\item{grayscale}{Defines input layer color channels -  `1` if `TRUE`, `3` if `FALSE`.}

\item{scale}{Scaling factor for images pixel values. Default to `1 / 255`.}

\item{anchors}{Prediction anchors. For exact format check \code{\link[platypus]{coco_anchors}}.}

\item{labels}{Character vector containing class labels. For example \code{\link[platypus]{coco_labels}}.}

\item{batch_size}{Batch size.}

\item{shuffle}{Should data be shuffled.}
}
\description{
Generates batches of data (images and box coordinates/scores). The data will be looped over (in batches).
}
