% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_loss.R
\name{yolo3_loss}
\alias{yolo3_loss}
\title{Generates `Yolo3` loss function.}
\usage{
yolo3_loss(
  anchors,
  n_class,
  nonobj_threshold = 0.5,
  bbox_lambda = 1,
  obj_lambda = 1,
  noobj_lambda = 1,
  class_lambda = 1,
  class_weights = rep(1, n_class)
)
}
\arguments{
\item{anchors}{Prediction anchors. For exact format check \code{\link[platypus]{coco_anchors}}.}

\item{n_class}{Number of prediction classes.}

\item{nonobj_threshold}{Non-object ignore threshold.}

\item{bbox_lambda}{Bounding box loss lambda.}

\item{obj_lambda}{Object loss lambda.}

\item{noobj_lambda}{Nonobject loss lambda.}

\item{class_lambda}{Class loss lambda.}

\item{class_weights}{Vector of length `n_class` with class weights.}
}
\value{
`Yolo3` loss function.
}
\description{
Generates `Yolo3` loss function.
}
