% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_loss.R
\name{yolo3_metrics}
\alias{yolo3_metrics}
\title{Generates `Yolo3` IoU metric function.}
\usage{
yolo3_metrics(anchors, n_class)
}
\arguments{
\item{anchors}{Prediction anchors. For exact format check \code{\link[platypus]{coco_anchors}}.}

\item{n_class}{Number of prediction classes.}
}
\value{
`Yolo3` IoU metric function.
}
\description{
Generates `Yolo3` IoU metric function.
}
