% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3.R
\name{yolo3_output}
\alias{yolo3_output}
\title{Creates Yolo3 output grid.}
\usage{
yolo3_output(inputs, filters, anchors_per_grid, n_class, name)
}
\arguments{
\item{inputs}{Models or layer objects.}

\item{filters}{Integer, the dimensionality of the output space (i.e. the number of output filters in the convolution).}

\item{anchors_per_grid}{Number of anchors/boxes per one output grid.}

\item{n_class}{Number of prediction classes.}

\item{name}{Model name.}
}
\value{
Yolo3 output grid.
}
\description{
Creates Yolo3 output grid of dimensionality `(S, H, W, anchors_per_grid, 5 + n_class)`.
}
