% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_fit.R
\name{yolo3_predict_generator}
\alias{yolo3_predict_generator}
\title{Calculates predictions on new samples using `YOLOv3` data generator.}
\usage{
yolo3_predict_generator(model, generator, steps)
}
\arguments{
\item{model}{`YOLOv3` Model.}

\item{generator}{`YOLOv3` Data generator.}

\item{steps}{Steps in epoch.}
}
\description{
Calculates predictions on new samples using `YOLOv3` data generator.
}
