% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_train_on_batch.R
\name{custom_fit_generator}
\alias{custom_fit_generator}
\title{Fits model using data generator.}
\usage{
custom_fit_generator(
  metric_names,
  model,
  generator,
  epochs,
  steps_per_epoch,
  validation_generator = NULL,
  validation_steps_per_epoch = NULL,
  model_filepath = NULL
)
}
\arguments{
\item{metric_names}{Metric names.}

\item{model}{Model.}

\item{generator}{Data generator.}

\item{epochs}{Number of epochs.}

\item{steps_per_epoch}{Steps per epoch.}

\item{validation_generator}{Validation data generator.}

\item{validation_steps_per_epoch}{Validation steps per epoch.}

\item{model_filepath}{Path to save the model.}
}
\description{
Fits model using data generator.
}
