% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.platypus.R
\name{cv.platypus}
\alias{cv.platypus}
\title{k-fold cross validation for platypus}
\usage{
cv.platypus(view.list, fn.labs, classcol.labs = 1, cv.folds = 10,
  n.iters = 100, majority.threshold.percent = 100,
  expanded.output = FALSE, updating = FALSE,
  ignore.label = "intermediate", parallel = FALSE,
  output.folder = NA)
}
\arguments{
\item{view.list}{List of view objects}

\item{fn.labs}{File containing outcome labels}

\item{classcol.labs}{Which column from the labels file to use for learning}

\item{cv.folds}{number of folds for label learning validation (similar to cross validation folds), default=10}

\item{n.iters}{Maximal number of iterations for each platypus run, default=100}

\item{majority.threshold.percent}{Percent agreement required to learn a sample's class label, default=100}

\item{expanded.output}{Expanded output: returned result list contains a list of trained views after each iteration, default=FALSE}

\item{updating}{Updating the accuracies of the single views in each iteration, default=FALSE}

\item{ignore.label}{Label class to ignore, if any. Defaults to 'intermediate'}

\item{parallel}{Whether or not to run in parallel mode.}

\item{output.folder}{Name of the folder where output is stored.}
}
\value{
A list containing fold.accuracy, labelling.matrix,labelling.matrices.views
}
\description{
k-fold cross validation for platypus
}
\keyword{platypus}
