% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_generator.R
\name{find_anchors_and_grids_for_true_boxes}
\alias{find_anchors_and_grids_for_true_boxes}
\title{Finds best anchors and output grid coordinates for true bounding box coordinates.}
\usage{
find_anchors_and_grids_for_true_boxes(true_boxes, anchors, true_grid)
}
\arguments{
\item{true_boxes}{True bounding box coordinates.}

\item{anchors}{Prediction anchors. For exact format check \code{\link[platypus]{coco_anchors}}.}

\item{true_grid}{`Yolo3` output grids.}
}
\value{
`data.frame` with best anchors and output grid coordinates for true bounding box coordinates.
}
\description{
Finds best anchors and output grid coordinates for true bounding box coordinates.
}
