% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_anchors.R
\name{generate_anchors}
\alias{generate_anchors}
\title{Calculates anchor boxes using k-mean++ algorithm.}
\usage{
generate_anchors(
  anchors_per_grid,
  annot_path,
  labels,
  n_iter = 10,
  annot_format = "pascal_voc",
  seed = 1234,
  centroid_fun = mean
)
}
\arguments{
\item{anchors_per_grid}{Number of anchors per one grid.}

\item{annot_path}{Annotations directory.}

\item{labels}{Character vector containing class labels. For example \code{\link[platypus]{coco_labels}}.}

\item{n_iter}{Maximum number of iteration for k-mean++ algorithm.}

\item{annot_format}{Annotations format. One of `pascal_voc`, `labelme`.}

\item{seed}{Random seed.}

\item{centroid_fun}{Function to use for centroid calculation.}
}
\value{
List of anchor boxes.
}
\description{
Calculates anchor boxes using k-mean++ algorithm.
}
