% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEX_GSEA.R
\name{GEX_GSEA}
\alias{GEX_GSEA}
\title{Conducts a Gene Set Enrichment Analysis (GSEA) on a set of genes submitted in a data frame with a metric each. Works with the output of GEX_genes_cluster or a custom data frame contianing the gene symbols either in a column "symbols" or as rownames and a metric for each gene. The name of the column containing the metric has to be declared via the input metric_colname.}
\usage{
GEX_GSEA(
  GEX.cluster.genes.output,
  MT.Rb.filter,
  filter,
  path_to_pathways,
  metric_colname,
  pval_adj_cutoff,
  Enrichment.Plots
)
}
\arguments{
\item{GEX.cluster.genes.output}{Data frame containing the list of gene symbols and a metric. Function works directly with GEX_cluster_genes output.}

\item{MT.Rb.filter}{Logical, should Mitotic and Ribosomal genes be filtered out of the geneset. True by default.}

\item{filter}{Character vector containing the identifying symbol sequence for the genes which should be filtered out, if MT.Rb.filter == T. By default set to c("MT-", "RPL", "RPS").}

\item{path_to_pathways}{Path to gmt file containing the gene sets. Can be downloaded from MSigDB.}

\item{metric_colname}{Name of column which contains the metric used for the ranking of the submitted genelist. "avg_logFC" is used by default.}

\item{pval_adj_cutoff}{Only genes with a more significant adjusted pvalue are considered. Default: 0.001}

\item{Enrichment.Plots}{List of Gene-set names which should be plotted as Enrichment plots in addition to the top 10 Up and Downregulated Genesets.}
}
\value{
Returns a list containing a tibble with the gene sets and their enrichment scores and Enrichment plots. List element [[1]]: Dataframe with Genesets and statistics. [[2]]: Enrichment plots of top10 Up regulated genesets. [[3]]: Enrichment plots of top10 Down regulated genesets. [[4]]: Enrichment plots of submited gene-sets in parameter Enrichment.Plot.
}
\description{
Conducts a Gene Set Enrichment Analysis (GSEA) on a set of genes submitted in a data frame with a metric each. Works with the output of GEX_genes_cluster or a custom data frame contianing the gene symbols either in a column "symbols" or as rownames and a metric for each gene. The name of the column containing the metric has to be declared via the input metric_colname.
}
\examples{
\dontrun{
df <- GEX_cluster_genes(gex_combined[[1]])
output <- GEX_GSEA(GEX.cluster.genes.output =  df[[1]], MT.Rb.filter = T, path_to_pathways = "./c5.go.bp.v7.2.symbols.gmt")
cowplot::plot_grid(plotlist=output[[2]], ncol=2)
View(gex_gsea[[1]])
}
}
