% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEX_phenotype_per_clone.R
\name{GEX_phenotype_per_clone}
\alias{GEX_phenotype_per_clone}
\title{Integrates VDJ and gene expression libraries by providing cluster membership seq_per_vdj object and the index of the cell in the Seurat RNA-seq object.}
\usage{
GEX_phenotype_per_clone(seurat.object, clonotype.ids)
}
\arguments{
\item{seurat.object}{A single seurat object from automate_GEX function after labeling cell phenotypes using the GEX_phenotype function.}

\item{clonotype.ids}{Output from either VDJ_analyze or VDJ_clonotype functions. This list should correspond to a single GEX.list object, in which each list element in clonotype.list is found in the GEX.object. Furthermore, these repertoires should be found in the automate_GEX library.}
}
\value{
Returns a stacked barplot that visualizes the seurat cluster membership for different cell phenotypes.
}
\description{
Integrates VDJ and gene expression libraries by providing cluster membership seq_per_vdj object and the index of the cell in the Seurat RNA-seq object.
}
\examples{
\dontrun{
GEX_phenotype_per_clone_plot <- GEX_phenotype_per_clone(seurat.object = automate.gex.output[[1]], clonotype.ids= c(1,2,3,4,5))
}

}
