% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEX_scatter_coexpression.R
\name{GEX_scatter_coexpression}
\alias{GEX_scatter_coexpression}
\title{Clonal frequency plot displaying clonal expansion for either T and B cells with Platypus v3 input.}
\usage{
GEX_scatter_coexpression(GEX, gene.1, gene.2, color.theme)
}
\arguments{
\item{GEX}{GEX seurat object generated with VDJ_GEX_matrix}

\item{gene.1}{Character. Name of a gene in rownames(VDJ.matrix)}

\item{gene.2}{Character. Name of a gene in rownames(VDJ.matrix)}

\item{color.theme}{Character. A color to use for the composite plot}
}
\value{
Returns a gridplot showing coexpression of gene.1 and gene.2
}
\description{
Clonal frequency plot displaying clonal expansion for either T and B cells with Platypus v3 input.
}
\examples{
\dontrun{
gene1 <- "BCL6"
gene2 <- "CD40"
ggsave(filename = paste0("Coex_", gene1,"_", gene2,"_scat.png"),
plot = GEX_scatter_coexpression(VDJ_comb[[2]], gene1,gene2),
dpi = 300, width = 12, height = 12)
}
}
