% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platypus.R
\name{platypus}
\alias{platypus}
\title{platypus multiview learning}
\usage{
platypus(fn.labs, view.list, ignore.label = "intermediate", i = 100,
  m = 100, u = FALSE, e = FALSE, updating = FALSE,
  expanded.output = FALSE)
}
\arguments{
\item{fn.labs}{File containing outcome labels}

\item{view.list}{List of view objects}

\item{i}{Maximal number of iterations for each platypus run, default=100}

\item{m}{Percent agreement required to learn a sample's class label, default=100}

\item{u}{Updating the accuracies of the single views in each iteration, default=FALSE}

\item{e}{Expanded output: returned result list contains a list of trained views after each iteration, default=FALSE}

\item{w}{flag for weighting the preditions by accuracy, default=FALSE # TODO REMOVE}

\item{b}{Label class to ignore, if any. Defaults to 'intermediate'}
}
\value{
final.result.list
}
\description{
The standard platypus function for multiview learning
first pass the filepath for the labs file and the column name or number of the class (if not given, first column is default)
pass the filepath of a parameter-file for each view
}
\keyword{platypus}
