% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_boxes}
\alias{plot_boxes}
\title{Generates raster images with bounding boxes.}
\usage{
plot_boxes(
  images_paths,
  boxes,
  labels,
  correct_hw = TRUE,
  target_size = NULL,
  grayscale = FALSE
)
}
\arguments{
\item{images_paths}{Image filepaths.}

\item{boxes}{List of `data.frames` with bounding boxes corresponding to the images.}

\item{labels}{Character vector containing class labels. For example \code{\link[platypus]{coco_labels}}.}

\item{correct_hw}{Logical. Should height/width rescaling of bounding boxes be applied.}

\item{target_size}{Image target size.}

\item{grayscale}{Should images be plotted in grayscale.}
}
\value{
Raster images with bounding boxes.
}
\description{
Generates raster images with bounding boxes.
}
