% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_boxes_ggplot}
\alias{plot_boxes_ggplot}
\title{Generates raster image with bounding boxes.}
\usage{
plot_boxes_ggplot(
  image_path,
  boxes,
  labels,
  correct_hw,
  target_size,
  grayscale
)
}
\arguments{
\item{image_path}{Image filepath.}

\item{boxes}{`data.frame` with bounding boxes corresponding to the image.}

\item{labels}{Character vector containing class labels. For example \code{\link[platypus]{coco_labels}}.}

\item{correct_hw}{Logical. Should height/width rescaling of bounding boxes be applied.}

\item{target_size}{Image target size.}

\item{grayscale}{Should images be plotted in grayscale.}
}
\value{
Raster image with bounding boxes.
}
\description{
Generates raster image with bounding boxes.
}
