% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_masks}
\alias{plot_masks}
\title{Generates raster images with segmentation masks.}
\usage{
plot_masks(
  images_paths,
  masks,
  labels,
  colormap,
  grayscale = FALSE,
  save_dir = NULL,
  plot_images = TRUE,
  background_label = "background"
)
}
\arguments{
\item{images_paths}{Image filepaths.}

\item{masks}{Predicted segmentation masks.}

\item{labels}{Character vector containing class labels. For example \code{\link[platypus]{voc_labels}}.}

\item{colormap}{Class color map. For example \code{\link[platypus]{voc_colormap}}.}

\item{grayscale}{Should images be plotted in grayscale.}

\item{save_dir}{Directory in which to save generated images.}

\item{plot_images}{Should images be plotted.}

\item{background_label}{Name of the background class label to filter out.}
}
\value{
Raster images with segmentation masks.
}
\description{
Generates raster images with segmentation masks.
}
