% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_generator.R
\name{segmentation_generator}
\alias{segmentation_generator}
\title{Generates batches of data (images and masks). The data will be looped over (in batches).}
\usage{
segmentation_generator(path, mode = "dir", classes = 2,
  only_images = FALSE, target_size = c(256, 256), grayscale = FALSE,
  scale = 1/255, batch_size = 32, shuffle = TRUE,
  subdirs = c("images", "masks"), column_sep = ";")
}
\arguments{
\item{path}{Images and masks directory.}

\item{mode}{Character. One of `c("dir", "nested_dirs", "config_file")`}

\item{classes}{Number of classes. Minimum is `2` (background + other object).}

\item{only_images}{Should generator read only images (e.g. on train set for predictions).}

\item{target_size}{Images / mask size (height, width). Default to `c(256, 256)`.}

\item{grayscale}{Boolean, whether to load the image as grayscale.}

\item{scale}{Scaling factor for images pixel values. Default to `1 / 255`.}

\item{batch_size}{Batch size.}

\item{shuffle}{Should data be shuffled.}

\item{subdirs}{Vector of two characters containing names of subdirectories with images and masks.}

\item{column_sep}{Character. Configuration file separator.}
}
\description{
Generates batches of data (images and masks). The data will be looped over (in batches).
}
