% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{training_history}
\alias{training_history}
\title{The epoch-by-epoch record}
\usage{
training_history(object)
}
\arguments{
\item{object}{A [platypus_fit()].}
}
\value{
A data frame with one row per model per epoch.
}
\description{
The epoch-by-epoch record
}
