% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_predict.R
\name{transform_boxes}
\alias{transform_boxes}
\title{Transforms `Yolo3` predictions into valid box coordinates/scores.}
\usage{
transform_boxes(
  preds,
  anchors,
  n_class,
  anchors_per_grid,
  obj_threshold,
  net_h,
  net_w
)
}
\arguments{
\item{preds}{\code{\link[platypus]{yolo3}} model predictions.}

\item{anchors}{Prediction anchors. For exact format check \code{\link[platypus]{coco_anchors}}.}

\item{n_class}{Number of prediction classes.}

\item{anchors_per_grid}{Number of anchors/boxes per one output grid.}

\item{obj_threshold}{Minimum objectness score. Must be in range `[0, 1]`. All boxes with objectness score less than `obj_threshold` will be filtered out.}

\item{net_h}{Input layer height from trained \code{\link[platypus]{yolo3}} model. Must be divisible by `32`.}

\item{net_w}{Input layer width from trained \code{\link[platypus]{yolo3}} model. Must be divisible by `32`.}
}
\value{
List of box coordinates/scores.
}
\description{
Transforms `Yolo3` predictions into valid box coordinates/scores.
}
