% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_loss.R
\name{transform_boxes_tf}
\alias{transform_boxes_tf}
\title{Transforms `Yolo3` predictions into valid box coordinates/scores.}
\usage{
transform_boxes_tf(preds, anchors, n_class, transform_proba = TRUE)
}
\arguments{
\item{preds}{\code{\link[platypus]{yolo3}} model predictions (from one grid).}

\item{anchors}{Prediction anchors (for one grid). For exact format check \code{\link[platypus]{coco_anchors}}.}

\item{n_class}{Number of prediction classes.}

\item{transform_proba}{Logical. Should the score/class probabilities be transformed.}
}
\value{
Transformed bounding box coordinates/scores.
}
\description{
Transforms `Yolo3` predictions into valid box coordinates/scores.
}
