% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_net.R
\name{u_net}
\alias{u_net}
\title{Creates a U-Net architecture.}
\usage{
u_net(input_img, filters, dropout = 0.1, batch_normalization = TRUE)
}
\arguments{
\item{input_img}{Model or layer object}

\item{filters}{Integer, the dimensionality of the output space (i.e. the number of output filters in the convolution).}

\item{dropout}{Dropout rate.}

\item{batch_normalization}{Shoud batch normalization be used in the block.}
}
\value{
Downsalmling U-Net block
}
\description{
Creates a U-Net architecture.
}
