% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_net.R
\name{u_net}
\alias{u_net}
\title{Creates a U-Net architecture.}
\usage{
u_net(
  net_h,
  net_w,
  grayscale,
  blocks = 4,
  n_class = 2,
  filters = 16,
  dropout = 0.1,
  batch_normalization = TRUE,
  kernel_initializer = "he_normal"
)
}
\arguments{
\item{net_h}{Input layer height. Must be equal to `2^x, x - natural`..}

\item{net_w}{Input layer width. Must be equal to `2^x, x - natural`.}

\item{grayscale}{Defines input layer color channels -  `1` if `TRUE`, `3` if `FALSE`.}

\item{blocks}{Number of blocks in the model.}

\item{n_class}{Number of classes. Minimum is `2` (background + other object).}

\item{filters}{Integer, the dimensionality of the output space (i.e. the number of output filters in the convolution).}

\item{dropout}{Dropout rate.}

\item{batch_normalization}{Should batch normalization be used in the block.}

\item{kernel_initializer}{Initializer for the kernel weights matrix.}
}
\value{
U-Net model.
}
\description{
Creates a U-Net architecture.
}
