% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ_clonal_lineages.R
\name{VDJ_clonal_lineages}
\alias{VDJ_clonal_lineages}
\title{Organizes and extracts full-length sequences for clonal lineage inference. The output sequence can either contain the germline sequence as determined by cellranger or can just contain the sequences contained in each clonal family.}
\usage{
VDJ_clonal_lineages(
  call_MIXCR.output,
  VDJ_extract_germline.output,
  as.nucleotide,
  with.germline,
  VDJ_GEX_matrix,
  platypus.version
)
}
\arguments{
\item{call_MIXCR.output}{The output from the call_MIXCR function containing the full-length VDJRegion sequences.}

\item{VDJ_extract_germline.output}{The output from the VDJ_extract_germline function. This should have the germline information. This needs to be supplied if the with.germline argument is set to true.}

\item{as.nucleotide}{Logical determining whether the full-length VDJRegion sequence should use nucleotide seqeunce. TRUE indicates nucleotide sequences and FALSE will extract amino acid sequences.}

\item{with.germline}{Logical determining whether the germline sequence as determined by cellranger should be included in the output list of sequences. If so, the germline will be added to the last row of each dataframe object.}

\item{platypus.version}{Default is "v2" for compatibility. To use the output of VDJ_GEX_matrix function, one should change this argument to "v3".}

\item{VDJ.GEX.matrix}{Output from the VDJ.GEX.matrix function. The output object should have the VDJ information (e.g., the original VDJ_GEX_matrix call should have had cellranger's VDJ output supplied as input).}
}
\value{
returns a list containing the sequences for each clonal family as determined by the input clonotyping strategy to call_MIXCR and VDJ_extract_germline. The outer list corresponds to distinct repertoires supplied to the call_MIXCR function (e.g. VDJ.clonal.lineage.output[[i]][[j]] will contain a dataframe of the j'th clone in the i'th repertoire)
}
\description{
Organizes and extracts full-length sequences for clonal lineage inference. The output sequence can either contain the germline sequence as determined by cellranger or can just contain the sequences contained in each clonal family.
}
\examples{
\dontrun{
clonal_lineages <- VDJ_clonal_lineages(call_MIXCR.output=call_MIXCR_output, VDJ_extract_germline.output=VDJ_extract_germline_output,as.nucleotide=F,with.germline=T)
}

}
