% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ_GEX_clonotype_clusters_circos.R
\name{VDJ_clonotype_clusters_circos}
\alias{VDJ_clonotype_clusters_circos}
\title{Makes a Circos plot from the VDJ_GEX_integrate output. Connects the clonotypes with the corresponding clusters.}
\usage{
VDJ_clonotype_clusters_circos(
  vdj.gex.integrate.output,
  topX,
  label.threshold,
  axis,
  c.count,
  n_cluster
)
}
\arguments{
\item{vdj.gex.integrate.output}{The output of the VDJ_GEX_integrate function. A list of data frames for each sample containing the clonotype information and cluster membership information.}

\item{label.threshold}{Minimal amount of clonotypes per gene neccessary to add a gene label to the sector. Default: 0.}

\item{c.count}{Show clonotype or cell count on Circos plot. Default = T.}

\item{TopX}{Plots only the top X most expanded clonotypes. By default all clonotypes are shown.}
}
\value{
Returns list of plots. The first n elements contain the circos plot of the n datasets from the VDJ.analyze function. The n+1 element contains a list of the n adjancey matrices for each dataset.
}
\description{
Makes a Circos plot from the VDJ_GEX_integrate output. Connects the clonotypes with the corresponding clusters.
}
\examples{
\dontrun{
 plots <- VDJ_clonotype_clusters_circos(vdj_gex_integrate_test, topX=100, label.threshold=5)
}
}
