% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ_isotypes_per_clone.R
\name{VDJ_isotypes_per_clone}
\alias{VDJ_isotypes_per_clone}
\title{Clonal frequency plot displaying the isotype usage of each clone.}
\usage{
VDJ_isotypes_per_clone(VDJ_clonotype_output, VDJ_per_clone_output, clones)
}
\arguments{
\item{VDJ_clonotype_output}{list of dataframes based on the VDJ_clonotype function output.}

\item{VDJ_per_clone_output}{list of dataframes based on the VDJ_per_clone function output.}

\item{clones}{numeric value indicating the number of clones to be displayed on the clonal expansion plot. Can take values between 1-50. Default value is 50.}
}
\value{
returns a list containing plots with the percentages of isotypes for each clone on the cell level.
}
\description{
Clonal frequency plot displaying the isotype usage of each clone.
}
\examples{
\dontrun{
VDJ.isotype.per.clone <- VDJ_isotypes_per_clone(VDJ_clonotype_output = VDJ.analyze.output, VDJ_per_clone_output = VDJ.per.clone.output, clones = 30)
}
}
