% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ_logoplot.R
\name{VDJ_logoplot}
\alias{VDJ_logoplot}
\title{Plots a logoplot of the CDR3 region}
\usage{
VDJ_logoplot(VDJ.object, length_cdr3)
}
\arguments{
\item{length_cdr3}{Integer indicating the length of the CDR3 regions that are selected to be plotted.}

\item{GEX.object}{Output of the automate_GEX function}
}
\description{
Plots a logoplot of the CDR3 region
}
\examples{
\dontrun{
GEX_fraction_cells_per_cluster(automate_GEX.output, per_sample))
}
}
