% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ_overlap_heatmap.R
\name{VDJ_overlap_heatmap}
\alias{VDJ_overlap_heatmap}
\title{Produces clonal overlap plots using output of VDJ_GEX_matrix}
\usage{
VDJ_overlap_heatmap(
  VDJ.matrix.output,
  feature.columns,
  grouping.column,
  pvalues.label.size,
  axis.label.size,
  add.barcode.table,
  platypus.version
)
}
\arguments{
\item{feature.columns}{A character array of column names of which the overlap should be displayed. The content of these columns is pasted together (separated by "/"). E.g. if the overlap in cells germline gene usage is desired, the input could be c("VDJ_jgene","VDJ_dgene","VDJ_vgene"). These columns would be pasted and compared across the grouping variable.}

\item{grouping.column}{A column which acts as a grouping variable. If repertoires are to be compared use the sample_id column.}

\item{pvalues.label.size}{Numeric. Defaults to 4. Is passed on to ggplot theme}

\item{axis.label.size}{Numeric. Defaults to 4. Is passed on to ggplot theme}

\item{add.barcode.table}{Boolean. Defaults to T. Whether to generate a dataframe with frequencies and barcodes of cells with overlapping features. This is useful to e.g. analyze deferentially expressed genes between cells of two samples or groups expressing the same VDJ or VJ chain}

\item{platypus.version}{Character. At the moment this function runs only on the output of the VDJ_GEX_matrix function meaning that it is exclusively part of Platypus "v3". With further updates the functionality will be extended.}

\item{VDJ.matrix}{VDJ matrix from the VDJ_GEX_matrix function}
}
\value{
A list of a ggplot (out[[1]]) and the table from which the ggplot was generated (out[[2]])
}
\description{
Produces clonal overlap plots using output of VDJ_GEX_matrix
}
\examples{
\dontrun{
#To test the overlap of barcodes between multiple samples
barcode_overlap <- VDJ_overlap_heatmap(VDJ.matrix = VDJ_GEX_matrix_output[[1]], feature.columns = c("orig_barcode") ,grouping.column = "repertoire", axis.label.size = 15)
}
}
