% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ_Vgene_usage.R
\name{VDJ_Vgene_usage}
\alias{VDJ_Vgene_usage}
\title{Produces a matrix counting the number of occurencies for each IgH and IgK/L Vgene combinations.}
\usage{
VDJ_Vgene_usage(VDJ.clonotype.output)
}
\arguments{
\item{VDJ.clonotype.output}{Output from VDJ_analyze function. This should be a list of clonotype dataframes, with each list element corresponding to a single VDJ repertoire.}
}
\value{
Returns a list of matrices containing the number of Vgene heavy/light chain combinations per repertoire.
}
\description{
Produces a matrix counting the number of occurencies for each IgH and IgK/L Vgene combinations.
}
\examples{
\dontrun{
example.vdj.vgene_usage <- VDJ_Vgene_usage(VDJ.clonotype.output = VDJ.clonotype.output)
}
}
\seealso{

}
