% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ_Vgene_usage_barplot.R
\name{VDJ_Vgene_usage_barplot}
\alias{VDJ_Vgene_usage_barplot}
\title{Produces a barplot with the most frequently used IgH and IgK/L Vgenes.}
\usage{
VDJ_Vgene_usage_barplot(
  clonotype.list,
  HC.gene.number,
  LC.Vgene,
  LC.gene.number
)
}
\arguments{
\item{clonotype.list}{Output from VDJ_analyze function. This should be a list of clonotype dataframes, with each list element corresponding to a single VDJ repertoire.}

\item{LC.Vgene}{Logical indicating whether to make a barplot of the LC V genes distribution. Default is set to FALSE.}

\item{LC.gene.number}{Numeric value indicating the top genes to be dispayed. If this number is higher than the total number of unique LC V genes in the VDJ repertoire, then this number is equal to the number of unique LC V genes.}

\item{HC, gene.number}{Numeric value indicating the top genes to be dispayed. If this number is higher than the total number of unique HC V genes in the VDJ repertoire, then this number is equal to the number of unique HC V genes.}
}
\value{
Returns a list of ggplot objects which show the distribution of IgH and IgK/L V genes for the most used V genes.
}
\description{
Produces a barplot with the most frequently used IgH and IgK/L Vgenes.
}
