% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3.R
\name{yolo3}
\alias{yolo3}
\title{Creates a Yolo3 architecture.}
\usage{
yolo3(
  net_h = 416,
  net_w = 416,
  grayscale = FALSE,
  n_class = 80,
  anchors = coco_anchors
)
}
\arguments{
\item{net_h}{Input layer height. Must be divisible by `32`.}

\item{net_w}{Input layer width. Must be divisible by `32`.}

\item{grayscale}{Defines input layer color channels -  `1` if `TRUE`, `3` if `FALSE`.}

\item{n_class}{Number of prediction classes.}

\item{anchors}{Prediction anchors. For exact format check \code{\link[platypus]{coco_anchors}}.}
}
\value{
Yolo3 model.
}
\description{
Creates a Yolo3 architecture.
}
