% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_fit.R
\name{yolo3_fit_generator}
\alias{yolo3_fit_generator}
\title{Fits `YOLOv3` model using data generator.}
\usage{
yolo3_fit_generator(model, generator, epochs, steps_per_epoch)
}
\arguments{
\item{model}{`YOLOv3` Model.}

\item{generator}{`YOLOv3` Data generator.}

\item{epochs}{Number of epochs.}

\item{steps_per_epoch}{Steps per epoch.}
}
\description{
Fits `YOLOv3` model using data generator.
}
