% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_loss.R
\name{yolo3_grid_loss}
\alias{yolo3_grid_loss}
\title{Calculates loss for one `Yolo3` grid.}
\usage{
yolo3_grid_loss(
  y_true,
  y_pred,
  anchors,
  n_class,
  net_h,
  net_w,
  nonobj_threshold
)
}
\arguments{
\item{y_true}{Tensor of true coordinates/scores.}

\item{y_pred}{Tensor of predicted coordinates/scores.}

\item{anchors}{Prediction anchors (for one grid). For exact format check \code{\link[platypus]{coco_anchors}}.}

\item{n_class}{Number of prediction classes.}

\item{net_h}{Input layer height from trained \code{\link[platypus]{yolo3}} model. Must be divisible by `32`.}

\item{net_w}{Input layer width from trained \code{\link[platypus]{yolo3}} model. Must be divisible by `32`.}

\item{nonobj_threshold}{Non-object ignore threshold.}
}
\value{
Loss for one `Yolo3` grid.
}
\description{
Calculates loss for one `Yolo3` grid.
}
