% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolo3_loss.R
\name{yolo3_loss}
\alias{yolo3_loss}
\title{Generates `Yolo3` loss function.}
\usage{
yolo3_loss(anchors, n_class, net_h, net_w, nonobj_threshold = 0.5)
}
\arguments{
\item{anchors}{Prediction anchors. For exact format check \code{\link[platypus]{coco_anchors}}.}

\item{n_class}{Number of prediction classes.}

\item{net_h}{Input layer height from trained \code{\link[platypus]{yolo3}} model. Must be divisible by `32`.}

\item{net_w}{Input layer width from trained \code{\link[platypus]{yolo3}} model. Must be divisible by `32`.}

\item{nonobj_threshold}{Non-object ignore threshold.}
}
\value{
`Yolo3` loss function.
}
\description{
Generates `Yolo3` loss function.
}
